% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{add_missinglabels_mar}
\alias{add_missinglabels_mar}
\title{Throw out labels at random}
\usage{
add_missinglabels_mar(df, formula = NULL, prob = 0.1)
}
\arguments{
\item{df}{data.frame; Data frame of interest}

\item{formula}{formula; Formula to indicate the outputs}

\item{prob}{numeric; Probability of removing the label}
}
\description{
Throw out labels at random
}

