% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LaplacianKernelLeastSquaresClassifier.R
\name{adjacency_knn}
\alias{adjacency_knn}
\title{Calculate knn adjacency matrix}
\usage{
adjacency_knn(X, distance = "euclidean", k = 6)
}
\arguments{
\item{X}{matrix; input matrix}

\item{distance}{character; distance metric used in the \code{dist} function}

\item{k}{integer; Number of neighbours}
}
\value{
Symmetric binary adjacency matrix
}
\description{
Calculates symmetric adjacency: objects are neighbours is either one of them is in the set of nearest neighbours of the other.
}
