% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Classifier.R
\name{BaseClassifier}
\alias{BaseClassifier}
\title{Classifier used for enabling shared documenting of parameters}
\usage{
BaseClassifier(X, y, X_u, verbose, scale, eps, x_center, intercept, lambda,
  y_scale, kernel, use_Xu_for_scaling, ...)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{verbose}{logical; Controls the verbosity of the output}

\item{scale}{logical; Should the features be normalized? (default: FALSE)}

\item{eps}{numeric; Stopping criterion for the maximinimization}

\item{x_center}{logical;  Should the features be centered?}

\item{intercept}{logical; Whether an intercept should be included}

\item{lambda}{numeric; L2 regularization parameter}

\item{y_scale}{logical; whether the target vector should be centered}

\item{kernel}{kernlab::kernel to use}

\item{use_Xu_for_scaling}{logical; whether the unlabeled objects should be used to determine the mean and scaling for the normalization}

\item{...}{Not used}
}
\description{
Classifier used for enabling shared documenting of parameters
}
