\docType{methods}
\name{boundaryplot}
\alias{boundaryplot}
\alias{boundaryplot,LeastSquaresClassifier-method}
\alias{boundaryplot,LinearSVM-method}
\alias{boundaryplot,LogisticLossClassifier-method}
\alias{boundaryplot,LogisticRegression-method}
\alias{boundaryplot,SGDSVM-method}
\title{Decision Boundary plot of a classifier}
\arguments{
  \item{object}{LeastSquaresClassifier object}

  \item{p}{ggplot object of classification problem
  generated by clplot}

  \item{object}{LogisticLossClassifier object}

  \item{p}{ggplot object of classification problem
  generated by clplot}

  \item{object}{SGDSVM object}

  \item{p}{ggplot object of classification problem
  generated by clplot}
}
\description{
  Decision Boundary plot of a classifier

  Boundary plot method for LogisticRegression

  Boundary plot method for LeastSquaresClassifier

  Boundary plot method for LogisticLossClassifier

  Boundary plot method for SGDSVM

  Predict class of new observations using a LinearSVM
}

