% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{c.CrossValidation}
\alias{c.CrossValidation}
\title{Merge result of cross-validation runs on single datasets into a the same object}
\usage{
\method{c}{CrossValidation}(...)
}
\arguments{
\item{...}{Named arguments for the different objects, where the name reflects the dataset name}
}
\description{
Merge result of cross-validation runs on single datasets into a the same object
}
