% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{clapply}
\alias{clapply}
\title{Use mclapply conditional on not being in RStudio}
\usage{
clapply(X, FUN, ..., mc.cores = getOption("mc.cores", 2L))
}
\arguments{
\item{X}{vector}

\item{FUN}{function to be applied to the elements of X}

\item{...}{optional arguments passed to FUN}

\item{mc.cores}{number of cores to use}
}
\description{
Use mclapply conditional on not being in RStudio
}
