% Generated by roxygen2 (4.0.2): do not edit by hand
\name{clplot}
\alias{clplot}
\title{Plot 2D classification problem with missing labels}
\usage{
clplot(X, y)
}
\arguments{
\item{X}{feature values for observations}

\item{y}{labels for observations}
}
\description{
Plot 2D classification problem with missing labels
}

