% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{cov_ml}
\alias{cov_ml}
\title{Maximum Likelihood Covariance estimate}
\usage{
cov_ml(X)
}
\arguments{
\item{X}{matrix with observations}
}
\description{
Maximum Likelihood Covariance estimate
}

