% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CPLeastSquaresClassifier}
\alias{CPLeastSquaresClassifier}
\title{Contrastive Pressimistic Least Squares Classifier}
\usage{
CPLeastSquaresClassifier(X, y, X_u = NULL, lambda1 = 0, lambda2 = 0,
  intercept = TRUE, x_center = TRUE, scale = FALSE, method = "default",
  max_iterations = 10000, alpha = 1000, eps = 1e-08, ...)
}
\arguments{
\item{X}{Design matrix, intercept term is added within the function}

\item{y}{Vector or factor with class assignments}

\item{X_u}{Design matrix of the unlabeled data, intercept term is added within the function}

\item{lambda1}{Regularization parameter in the unlabeled+labeled data regularized least squares}

\item{lambda2}{Regularization parameter in the labeled data only regularized least squares}

\item{intercept}{TRUE if an intercept should be added to the model}

\item{x_center}{logical; Whether the feature vectors should be centered}

\item{scale}{logical; If TRUE, apply a z-transform to all observations in X and X_u before running the regression}

\item{method}{Either "default" for solving using L-BFGS-B gradient descent or "QP" fpr a quadratic programming based solution}

\item{max_iterations}{integer; Number of iterations in minimax solving}

\item{alpha}{numeric; Learning rate in minimax solving}

\item{eps}{numeric; Convergence criretion in minimax solving}

\item{...}{additional arguments}
}
\value{
S4 object of class CPLeastSquaresClassifier with the following slots:
\item{theta}{weight vector}
\item{classnames}{the names of the classes}
\item{modelform}{formula object of the model used in regression}
\item{scaling}{a scaling object containing the paramters of the z-transforms applied to the data}
\item{optimization}{the object returned by the optim function}
\item{unlabels}{the labels assigned to the unlabeled objects}
}
\description{
Loog's Contrastive Pressimism applied to the Least Squares Classifier
}

