% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/CPLinearDiscriminantAnalysis.R
\name{CPLinearDiscriminantClassifier}
\alias{CPLinearDiscriminantClassifier}
\title{Constrastive Pessimistic Semi-Supervised Linear Discriminant Classifier}
\usage{
CPLinearDiscriminantClassifier(X, y, X_u, prior = NULL, scale = FALSE,
  init = NULL, max_iterations = 1000, alpha = 1, eps = 1e-04,
  x_center = FALSE, ...)
}
\arguments{
\item{X}{design matrix of the labeled objects}

\item{y}{vector with labels}

\item{X_u}{design matrix of the labeled objects}

\item{prior}{<what param does>}

\item{scale}{Should the features be normalized? (default: FALSE)}

\item{init}{not currently used}

\item{max_iterations}{Maximimum number of iterations for the maximinimization}

\item{alpha}{numeric; Learning rate in the minimax solving}

\item{eps}{Stopping criterion for the maximinimization}

\item{x_center}{logical; Whether the feature vectors should be centered}

\item{...}{Additional Parameters, Not used}
}
\description{
Marco Loog's CPLDA classifier
}

