% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CrossValidation}
\alias{CrossValidation}
\title{Regular cross-validation on a (set of) classifier(s)}
\usage{
CrossValidation(X, y, classifiers, measures = c("predict", "losstest"),
  groups = NULL, k = 10, repeats = 1, verbose = FALSE)
}
\arguments{
\item{X}{design matrix of the labeled objects}

\item{y}{vector with labels}

\item{classifiers}{list; Classifiers to crossvalidate}

\item{measures}{character; c("predict","losstest") giving the measures to be used}

\item{groups}{vector; Group assigments to be taken into account in defining the folds}

\item{k}{integer; Number of folds in the cross-validation}

\item{repeats}{integer; Number of repeated assignments to folds}

\item{verbose}{logical; Controls the verbosity of the output}
}
\description{
Regular cross-validation on a (set of) classifier(s)
}

