% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{CrossValidationSSL}
\alias{CrossValidationSSL}
\title{Semi-supervised cross-validation}
\usage{
CrossValidationSSL(X, y, classifiers, n_labeled = 100,
  measures = c("predict", "losstest"), groups = NULL, k = 2,
  repeats = 1, verbose = FALSE, prop_unlabeled = 0.5)
}
\arguments{
\item{X}{Design matrix}

\item{y}{label vector}

\item{classifiers}{list with classification functions}

\item{n_labeled}{number of labeled objects to use}

\item{measures}{which measures to use}

\item{groups}{TODO: description or drop}

\item{k}{number of folds}

\item{repeats}{number of repeats}

\item{verbose}{Boolean: verbose output or not}

\item{prop_unlabeled}{proportion of the non-labeled objects to be used as unlabeled objects vs. test objects}
}
\description{
Semi-supervised cross-validation
}
\details{
Labeled samples form the fold
}

