% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CrossValidationSSL2}
\alias{CrossValidationSSL2}
\title{Crossvalidation with unlabeled objects}
\usage{
CrossValidationSSL2(X, y, classifiers, n_labeled = 100, groups = NULL,
  k = 2, repeats = 1, n_min = 1, verbose = FALSE,
  dataset_name = "Unknown Dataset")
}
\arguments{
\item{X}{Design matrix}

\item{y}{label vector}

\item{classifiers}{list with classification functions}

\item{n_labeled}{number of labeled objects to use}

\item{groups}{TODO: unused?}

\item{k}{number of folds}

\item{repeats}{number of repeats}

\item{n_min}{Minimum number of objects per class}

\item{verbose}{Boolean: verbose output or not}

\item{dataset_name}{Name of the dataset}
}
\description{
The folds are taken over the test sets, not the labeled sets.
TODO: include inside the other method
}

