\name{CrossValidationTransductive}
\alias{CrossValidationTransductive}
\title{Crossvalidation in the transductive setting}
\usage{
CrossValidationTransductive(X, y, classifiers, measures = c("predict",
  "losstest"), groups = NULL, k = 10, repeats = 1, verbose = FALSE,
  dataset_name = "Unknown Dataset")
}
\description{
Crossvalidation in the transductive setting
}

