% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinearSVM.R
\docType{methods}
\name{decisionvalues,LinearSVM-method}
\alias{decisionvalues,LinearSVM-method}
\title{decisionvalues for LinearSVM}
\usage{
\S4method{decisionvalues}{LinearSVM}(object, newdata)
}
\description{
decisionvalues for LinearSVM
}

