% Generated by roxygen2 (4.0.2): do not edit by hand
\name{DensityAdaptiveRegression}
\alias{DensityAdaptiveRegression}
\title{Density Adaptive Semi-supervised Regression}
\usage{
DensityAdaptiveRegression(X, y, X_u, alpha = "CV", kernel = "gaussian")
}
\arguments{
\item{X}{<what param does>}

\item{y}{<what param does>}

\item{X_u}{<what param does>}

\item{alpha}{<what param does>}

\item{kernel}{<what param does>}
}
\description{
Implementation of the method by Singh, Nowak and Zhu
}

