\name{DensitySensitiveRegression}
\alias{DensitySensitiveRegression}
\title{Density Sentisitive Semi-supervised Regression}
\usage{
DensitySensitiveRegression(X, y, X_u, alpha = "CV", kernel = "gaussian")
}
\arguments{
  \item{X}{<what param does>}

  \item{y}{<what param does>}

  \item{X_u}{<what param does>}

  \item{alpha}{<what param does>}

  \item{kernel}{<what param does>}
}
\description{
Implementation of method by Azizyan, Singh and Wasserman
}

