% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LearningCurve.R
\name{DifferencePlot}
\alias{DifferencePlot}
\title{Plot of difference between supervised and semi-supervised errors}
\usage{
DifferencePlot(data, measurement = 1, legendsetting = "right",
  dataset_names = NULL, classifier_names = NULL)
}
\arguments{
\item{data}{LearningCurve object}

\item{measurement}{Index of the measurement to be plotted}

\item{legendsetting}{Where the legend should be placed}

\item{dataset_names}{Names of the datasets in the LearningCurve object}

\item{classifier_names}{Names of the classifiers in the LearningCurve object}
}
\description{
Plot of difference between supervised and semi-supervised errors
}

