\name{EMLinearDiscriminantClassifier}
\alias{EMLinearDiscriminantClassifier}
\title{Expectation Maximization Nearest Mean Classifier}
\usage{
EMLinearDiscriminantClassifier(X, y, X_u, method = "EM", scale = FALSE,
  eps = 1e-08, verbose = FALSE, ...)
}
\description{
Expectation Maximization Nearest Mean Classifier
}

