% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMNearestMeanClassifier.R
\name{EMNearestMeanClassifier}
\alias{EMNearestMeanClassifier}
\title{Semi-Supervised Nearest Mean Classifier using Expectation Maximization}
\usage{
EMNearestMeanClassifier(X, y, X_u, method = "EM", scale = FALSE,
  eps = 1e-04)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{method}{character; Currently only "EM"}

\item{scale}{Should the features be normalized? (default: FALSE)}

\item{eps}{Stopping criterion for the maximinimization}
}
\description{
Semi-Supervised Nearest Mean Classifier using Expectation Maximization
}

