% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ERLogisticLossClassifier.R
\name{ERLogisticLossClassifier}
\alias{ERLogisticLossClassifier}
\title{Entropy Regularized Logistic Loss Classifier}
\usage{
ERLogisticLossClassifier(X, y, X_u = NULL, lambda = 0, lambda_entropy = 1,
  intercept = TRUE, init = NA, scale = FALSE, x_center = FALSE)
}
\arguments{
\item{X}{Design matrix, intercept term is added within the function}

\item{y}{Vector or factor with class assignments}

\item{X_u}{Design matrix of the unlabeled data, intercept term is added within the function}

\item{lambda}{numeric; Regularization parameter for the supervised loss}

\item{lambda_entropy}{numeric; Parameter for the entropy term}

\item{intercept}{TRUE if an intercept should be added to the model}

\item{init}{Starting parameter vector for gradient descent}

\item{scale}{If TRUE, apply a z-transform to all observations in X and X_u before running the regression}

\item{x_center}{logical; whether the data should be centered}

\item{...}{additional arguments}
}
\description{
Entropy Regularized Logistic Loss Classifier
}

