\name{ErrorCurveSSL}
\alias{ErrorCurveSSL}
\title{Generate Semi-Supervised Learning error curve}
\usage{
ErrorCurveSSL(X, y, classifiers, n_l, with_replacement = FALSE,
  sizes = 2^(1:8), n_test = 1000, repeats = 100, verbose = FALSE,
  n_min = 1)
}
\description{
Generate Semi-Supervised Learning error curve
}

