% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Measures.R
\name{measure_accuracy}
\alias{measure_accuracy}
\alias{measure_error}
\alias{measure_losslab}
\alias{measure_losstest}
\alias{measure_losstrain}
\title{Performance measures used in classifier evaluation}
\usage{
measure_accuracy(trained_classifier, X_l, y_l, X_u, y_u, X_test, y_test)

measure_error(trained_classifier, X_l, y_l, X_u, y_u, X_test, y_test)

measure_losstest(trained_classifier, X_l, y_l, X_u, y_u, X_test, y_test)

measure_losslab(trained_classifier, X_l, y_l, X_u, y_u, X_test, y_test)

measure_losstrain(trained_classifier, X_l, y_l, X_u, y_u, X_test, y_test)
}
\arguments{
\item{trained_classifier}{the trained classifier object}

\item{X_l}{design matrix with labeled object}

\item{y_l}{labels of labeled objects}

\item{X_u}{design matrix with unlabeled object}

\item{y_u}{labels of unlabeled objects}

\item{X_test}{design matrix with test object}

\item{y_test}{labels of test objects}
}
\description{
Classification error on test set

Classification error on test set

Average loss on test objects

Average loss on labeled objects

Arevage loss on labeled and unlabeled objects
}
\details{
Performance measures that can be used in \code{\link{CrossValidationSSL}} and \code{\link{LearningCurveSSL}}
}
\section{Functions}{
\itemize{
\item \code{measure_error}: 

\item \code{measure_losstest}: 

\item \code{measure_losslab}: 

\item \code{measure_losstrain}: 
}}

