% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WellSVM.R
\name{find_a_violated_label}
\alias{find_a_violated_label}
\title{Find a violated label}
\usage{
find_a_violated_label(alpha, K, y, ind_y, lr, y_init)
}
\arguments{
\item{alpha}{classifier weights}

\item{K}{kernel matrix}

\item{y}{label vector}

\item{ind_y}{Labeled/Unlabeled indicator}

\item{lr}{positive ratio}

\item{y_init}{label initialization}
}
\description{
Find a violated label
}
