% Generated by roxygen2 (4.0.2): do not edit by hand
\name{Generate2ClassGaussian}
\alias{Generate2ClassGaussian}
\title{Generate data from 2 gaussian distributed datasets}
\usage{
Generate2ClassGaussian(n = 10000, d = 100, var = 1, expected = TRUE)
}
\arguments{
\item{n}{integer; Number of examples to generate}

\item{d}{integer; dimensionality of the problem}

\item{var}{numeric; size of the variance parameter}

\item{expected}{logical; whether the decision boundary should be the expected or perpendicular}
}
\description{
Generate data from 2 gaussian distributed datasets
}

