% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenerateSSLData.R
\name{generateFourClusters}
\alias{generateFourClusters}
\title{Generate Four Clusters dataset}
\usage{
generateFourClusters(n = 100, distance = 6, expected = FALSE)
}
\arguments{
\item{n}{number of observations to generate}

\item{distance}{Distance between clusters (default: 6)}

\item{expected}{TRUE if the large margin equals the class boundary, FALSE if the class boundary is perpendicular to the large margin}
}
\description{
Generate a four clusters dataset
}
\examples{
D <- generateFourClusters(1000,distance=6,expected=TRUE)
plot(data[,1],data[,2],col=data$Class,asp=1)
}

