% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSSLData.R
\name{generateParallelPlanes}
\alias{generateParallelPlanes}
\title{Generate Parallel planes}
\usage{
generateParallelPlanes(n = 100, classes = 3, sigma = 0.1)
}
\arguments{
\item{n}{Number of objects to generate}

\item{classes}{Number of classes}

\item{sigma}{Noise added}
}
\description{
Generate Parallel planes
}
\examples{
library(ggplot2)
df <- generateParallelPlanes(100,3)
ggplot(df, aes(x=x,y=y,color=Class,shape=Class)) +
 geom_point()
}

