% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSSLData.R
\name{generateSpirals}
\alias{generateSpirals}
\title{Generate Intersecting Spirals}
\usage{
generateSpirals(n = 100, sigma = 0.1)
}
\arguments{
\item{n}{Number of objects to generate per class}

\item{sigma}{Noise added}
}
\description{
Generate Intersecting Spirals
}
\examples{
data <- generateSpirals(100,sigma=0.1)
#plot3D::scatter3D(data$x,data$y,data$z,col="black")
}

