% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plotting.R
\name{geom_classifier}
\alias{geom_classifier}
\title{Plot RSSL classifier boundary}
\usage{
geom_classifier(..., show_guide = TRUE)
}
\arguments{
\item{...}{List of trained classifiers}

\item{show_guide}{logical (default: TRUE); Show legend}
}
\description{
Plot RSSL classifier boundary
}

