% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{geom_linearclassifier}
\alias{geom_linearclassifier}
\title{Plot RSSL classifier boundary}
\usage{
geom_linearclassifier(..., show_guide = TRUE)
}
\arguments{
\item{...}{List of trained classifiers}

\item{show_guide}{logical (default: TRUE); Show legend}
}
\description{
Plot RSSL classifier boundary
}

