% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Plotting.R
\name{geom_responsibilities}
\alias{geom_responsibilities}
\title{Plot RSSL responsibilities}
\usage{
geom_responsibilities(classifier, data)
}
\arguments{
\item{classifier}{Classifier; Classifier to plot the responsibilities for}

\item{data}{data.frame; Data corresponding to the coordinates of the unlabeled objects}
}
\description{
Plot RSSL responsibilities
}

