% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRFClassifier.R
\name{GRFClassifier}
\alias{GRFClassifier}
\title{Gaussian Random Fields and Harmonic functions}
\usage{
GRFClassifier(X, y, X_u, adjacency_kernel = NULL, sigma = 0.1, eta = 0.1,
  class_mass_normalization = TRUE, scale = FALSE, x_center = FALSE,
  y_u = NULL)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{adjacency_kernel}{kernlab::kernel; kernel object}

\item{sigma}{parameter for kernel if no kernel is given}

\item{eta}{The influence of the external classifier (not currently used)}

\item{class_mass_normalization}{Should the Class Mass Normalization heuristic be applied? (default: TRUE)}

\item{scale}{logical; Should the features be normalized? (default: FALSE)}

\item{x_center}{logical;  Should the features be centered?}

\item{y_u}{Labels given by external classifier (Not used)}
}
\description{
Gaussian Random Fields and Harmonic functions
}
\examples{
library(ggplot2)
library(kernlab)

testdata <- generateTwoCircles(200,noise=0.1)
testdata[-sample(1:nrow(testdata),10),]$Class <- NA
tvec<-na.omit(testdata$Class)
dmatU <- as.matrix(testdata[is.na(testdata$Class),1:2])
dmat <- as.matrix(testdata[!is.na(testdata$Class),1:2])

classifier <- GRFClassifier(dmat, tvec, dmatU, 
                adjacency_kernel = rbfdot(10))

plotframe <- data.frame(dmatU, 
                        Output=classifier@responsibilities[,1])

ggplot(plotframe,aes(x=X1,y=X2,color=Output)) +
  geom_point() + coord_equal()

data <- generateParallelPlanes()
data$Class <- NA
data$Class[1] <- "a"
data$Class[101] <- "b"
data$Class[201] <- "c"
data$Class <- factor(data$Class)

GRFClassifier(Class~.,data,NULL)@responsibilities

data <- generateSpirals()
data$Class<-NA
data$Class[1] <- "a"
data$Class[200] <- "b"
data$Class <- factor(data$Class)

GRFClassifier(Class~.,data,NULL)@responsibilities
}
\references{
Zhu, X., Ghahramani, Z. & Lafferty, J., 2003. Semi-supervised learning using gaussian fields and harmonic functions. In Proceedings of the 20th International Conference on Machine Learning. pp. 912-919.
}

