% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GRFClassifier.R
\name{harmonic_function}
\alias{harmonic_function}
\title{Direct R Translation of Xiaojin Zhu's Matlab code to determine harmonic solution}
\usage{
harmonic_function(W, Y)
}
\arguments{
\item{W}{matrix; weight matrix where the fist L rows/column correspond to the labeled examples.}

\item{Y}{matrix; l by c 0,1 matrix encoding class assignments for the labeled objects}
}
\value{
The harmonic solution, i.e. eq (5) in the ICML paper, with or without class mass normalization
}
\description{
Direct R Translation of Xiaojin Zhu's Matlab code to determine harmonic solution
}
