\name{ICLeastSquaresRegression}
\alias{ICLeastSquaresRegression}
\title{<brief desc>}
\usage{
ICLeastSquaresRegression(X, y, X_u, lambda1 = 0, lambda2 = 0,
  scale = FALSE, lowerbound = -Inf, upperbound = Inf, ...,
  method = "default")
}
\arguments{
  \item{X}{<what param does>}

  \item{y}{<what param does>}

  \item{X_u}{<what param does>}

  \item{lambda1}{<what param does>}

  \item{lambda2}{<what param does>}

  \item{scale}{<what param does>}

  \item{lowerbound}{<what param does>}

  \item{upperbound}{<what param does>}

  \item{...}{<what param does>}

  \item{method}{<what param does>}
}
\description{
<full description>
}

