% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ICLinearDiscriminantClassifier.R
\name{ICLinearDiscriminantClassifier}
\alias{ICLinearDiscriminantClassifier}
\title{Implicitly Constrained Semi-Supervised Linear Discriminant Classifier}
\usage{
ICLinearDiscriminantClassifier(X, y, X_u, prior = NULL, scale = FALSE,
  init = NULL, sup_prior = FALSE, factr = 1e+07, x_center = FALSE, ...)
}
\arguments{
\item{X}{design matrix of the labeled objects}

\item{y}{vector with labels}

\item{X_u}{design matrix of the labeled objects}

\item{prior}{set a fixed class prior}

\item{scale}{logical; Should the features be normalized? (default: FALSE)}

\item{init}{not currently used}

\item{sup_prior}{logical; use the prior estimates based only on the labeled data, not the imputed labels (default: FALSE)}

\item{factr}{change the precision requirement passed to the optim function (default: 1e7)}

\item{x_center}{logical; Whether the data should be centered}

\item{...}{Additional Parameters, Not used}
}
\description{
Semi-supervised version of Linear Discriminant Analysis using implicit constraints.
This method finds the labeling of the unlabeled objects, whose resulting LDA solution gives the highest log-likelihood when evaluated on the labeled objects only.
}

