% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ICLogisticLossClassifier}
\alias{ICLogisticLossClassifier}
\title{Implicitly Constained Logistic Loss Classifier}
\usage{
ICLogisticLossClassifier(X, y, X_u, lambda1 = 0, lambda2 = 0,
  intercept = TRUE, scale = FALSE, init = NA, verbose = TRUE,
  lambda3 = 0, trueprob = NULL, ...)
}
\arguments{
\item{X}{Design matrix, intercept term is added within the function}

\item{y}{Vector or factor with class assignments}

\item{X_u}{Design matrix of the unlabeled data, intercept term is added within the function}

\item{lambda1}{Regularization parameter in the unlabeled+labeled data regularized least squares}

\item{lambda2}{Regularization parameter in the labeled data only regularized least squares}

\item{intercept}{TRUE if an intercept should be added to the model}

\item{scale}{If TRUE, apply a z-transform to all observations in X and X_u before running the regression}

\item{init}{Starting parameter vector for gradient descent}

\item{verbose}{logical; Verbose output}

\item{lambda3}{numeric; Prior controlling the deviation of mean predictions form the overall mean}

\item{trueprob}{numeric; Set overal mean}

\item{...}{additional arguments}
}
\value{
S4 object of class ICLogisticLossClassifier with the following slots:
\item{theta}{weight vector}
\item{classnames}{the names of the classes}
\item{modelform}{formula object of the model used in regression}
\item{scaling}{a scaling object containing the paramters of the z-transforms applied to the data}
\item{optimization}{the object returned by the optim function}
\item{unlabels}{the labels assigned to the unlabeled objects}
}
\description{
Implicitly constrained semisupervised learning with logistic loss. This is equivalent to finding a weighting of the unlabeled objects belonging to either of the two classes.
}

