% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LaplacianSVM.R
\name{LaplacianSVM}
\alias{LaplacianSVM}
\title{SVM Classifier}
\usage{
LaplacianSVM(X, y, X_u = NULL, lambda = 1, gamma = 1, scale = TRUE,
  kernel = vanilladot(), adjacency_kernel = rbfdot(1/4))
}
\arguments{
\item{X}{Design matrix}

\item{y}{Labels of the observations}

\item{scale}{Whether a z-transform should be applied (default: TRUE)}

\item{C}{Cost variable}
}
\value{
S4 object of type LaplacianSVM
}
\description{
SVM Classifier
}

