% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErrorCurve.R
\name{LearningCurve.matrix}
\alias{LearningCurve.matrix}
\title{Generate Learning Curve for a set of RSSL classifiers.}
\usage{
LearningCurve.matrix(X, y, classifiers, with_replacement = FALSE,
  sizes = seq(10, nrow(X), 10), n_test = 1000, repeats = 100,
  verbose = TRUE)
}
\arguments{
\item{X}{design matrix of the labeled objects}

\item{y}{vector with labels}

\item{classifiers}{list; Classifiers to crossvalidate}

\item{with_replacement}{logical; Whether objects should be drawn with replacement}

\item{sizes}{integer vector; sizes of the training sets}

\item{n_test}{integer; number of test objects to be drawn (only when with_replacement==TRUE)}

\item{repeats}{integer; Number of learning curves to average over}

\item{verbose}{logical; Controls verbosity of the output}
}
\value{
A list with the following items:
}
\description{
\code{LearningCurve} runs learning curve experiments for a given dataset and set of classifiers
}
\details{
This function generates the supervised learning curve for one or several classifiers. The learning curve is generated by increasing the number of training points used to train the classifier and evaluating performance.
}

