% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErrorCurve.R
\name{learningcurve_fraction_labeled}
\alias{learningcurve_fraction_labeled}
\title{Generate Learning Curve for losses at different fractions of objects labeled}
\usage{
learningcurve_fraction_labeled(X, y, classifiers, fracs = seq(0.1, 0.9, by =
  0.1), with_replacement = FALSE, repeats = 10, verbose = FALSE,
  n_min = 1, test_fraction = NULL)
}
\description{
Generate Learning Curve for losses at different fractions of objects labeled
}

