% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LeastSquaresClassifier.R
\name{LeastSquaresClassifier}
\alias{LeastSquaresClassifier}
\title{Least Squares Classifier}
\usage{
LeastSquaresClassifier(X, y, lambda = 0, intercept = TRUE,
  x_center = FALSE, scale = FALSE, method = "inverse", y_scale = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{lambda}{Regularization parameter of the l2 penalty}

\item{intercept}{TRUE if an intercept should be added to the model}

\item{x_center}{TRUE, whether the dependent variables (features) should be centered}

\item{scale}{If TRUE, apply a z-transform to the design matrix X before running the regression}

\item{method}{Method to use for fitting. One of c("inverse","Normal","QR","BFGS")}

\item{y_scale}{If True scale the target vector}
}
\value{
S4 object of class LeastSquaresClassifier with the following slots:
\item{theta}{weight vector}
\item{classnames}{the names of the classes}
\item{modelform}{formula object of the model used in regression}
\item{scaling}{a scaling object containing the paramters of the z-transforms applied to the data}
}
\description{
Use least squares regression as a classification technique using class indicators as targets. Note this method minimizes quadratic loss, not the truncated quadratic loss.
}

