% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/LeastSquaresClassifier.R,
%   R/NormalBasedClassifier.R, R/LogisticRegression.R, R/LinearSVM.R,
%   R/LogisticLossClassifier.R, R/QuadraticDiscriminantClassifier.R,
%   R/SelfLearning.R
\docType{methods}
\name{line_coefficients}
\alias{line_coefficients}
\alias{line_coefficients,LeastSquaresClassifier-method}
\alias{line_coefficients,NormalBasedClassifier-method}
\alias{line_coefficients,LogisticRegression-method}
\alias{line_coefficients,LinearSVM-method}
\alias{line_coefficients,LogisticLossClassifier-method}
\alias{line_coefficients,QuadraticDiscriminantClassifier-method}
\alias{line_coefficients,SelfLearning-method}
\title{Loss of a classifier or regression function}
\usage{
line_coefficients(object, ...)

\S4method{line_coefficients}{LeastSquaresClassifier}(object)

\S4method{line_coefficients}{NormalBasedClassifier}(object)

\S4method{line_coefficients}{LogisticRegression}(object)

\S4method{line_coefficients}{LinearSVM}(object)

\S4method{line_coefficients}{LogisticLossClassifier}(object)

\S4method{line_coefficients}{QuadraticDiscriminantClassifier}(object)

\S4method{line_coefficients}{SelfLearning}(object)
}
\arguments{
\item{object}{Classifier; Trained Classifier object}

\item{...}{Not used}
}
\value{
numeric of the total loss on the test data
}
\description{
Loss of a classifier or regression function
}
