% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearSVM.R
\name{LinearSVM}
\alias{LinearSVM}
\title{Linear SVM Classifier}
\usage{
LinearSVM(X, y, C = 1, method = "Dual", scale = TRUE, eps = 1e-09,
  reltol = 1e-13, maxit = 100)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{C}{Cost variable}

\item{method}{Estimation procedure c("Dual","Primal","BGD","SGD","Pegasos")}

\item{scale}{Whether a z-transform should be applied (default: TRUE)}

\item{eps}{Small value to ensure positive definiteness of the matrix in QP formulation}

\item{reltol}{relative tolerance using during BFGS optimization}

\item{maxit}{Maximum number of iterations for BFGS optimization}
}
\value{
S4 object of type LinearSVM
}
\description{
Linear SVM Classifier
}

