% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSVM.R
\name{LinearTSVM}
\alias{LinearTSVM}
\title{Linear CCCP Transductive SVM classifier}
\usage{
LinearTSVM(X, y, X_u, C, Cstar, s = 0, x_center = FALSE, scale = FALSE,
  eps = 1e-06, verbose = FALSE, init = NULL)
}
\arguments{
\item{X}{matrix; Design matrix, intercept term is added within the function}

\item{y}{vector; Vector or factor with class assignments}

\item{X_u}{matrix; Design matrix of the unlabeled data, intercept term is added within the function}

\item{C}{numeric; Cost parameter of the SVM}

\item{Cstar}{numeric; Cost parameter of the unlabeled objects}

\item{s}{numeric; parameter controlling the loss function of the unlabeled objects}

\item{x_center}{logical;  Should the features be centered?}

\item{scale}{logical; If TRUE, apply a z-transform to all observations in X and X_u before running the regression}

\item{eps}{numeric; Convergence criterion}

\item{verbose}{logical; print debugging messages (default: FALSE)}

\item{init}{numeric; Initial classifier parameters to start the convex concave procedure}
}
\description{
This method is mostly for debugging purposes, since its updates are done using numerical gradient calculations.
}

