% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4VM.R
\name{localDescent}
\alias{localDescent}
\title{Local descent}
\usage{
localDescent(instance, label, labelNum, unlabelNum, gamma, C, beta, alpha)
}
\arguments{
\item{instance}{Design matrix}

\item{label}{label vector}

\item{labelNum}{Number of labeled objects}

\item{unlabelNum}{Number of unlabeled objects}

\item{gamma}{Parameter for RBF kernel}

\item{C}{cost parameter for SVM}

\item{beta}{Controls fraction of objects assigned to positive class}

\item{alpha}{Controls fraction of objects assigned to positive class}
}
\value{
list(predictLabel=predictLabel,acc=acc,values=values,model=model)
}
\description{
Local descent used in S4VM
}
