% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{LogisticRegression}
\alias{LogisticRegression}
\title{(Regularized) Logistic Regression implementation used for comparison with semi-supervised implementations}
\usage{
LogisticRegression(X, y, lambda = 0, intercept = TRUE, scale = FALSE,
  init = NA, x_center = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{lambda}{numeric; L2 regularization parameter}

\item{intercept}{logical; Whether an intercept should be included}

\item{scale}{logical; Should the features be normalized? (default: FALSE)}

\item{init}{numeric; Initialization of parameters for the optimization}

\item{x_center}{logical;  Should the features be centered?}
}
\description{
(Regularized) Logistic Regression implementation used for comparison with semi-supervised implementations
}

