% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LogisticRegression.R
\name{LogisticRegressionFast}
\alias{LogisticRegressionFast}
\title{Logistic Regression implementation that uses R's glm}
\usage{
LogisticRegressionFast(X, y, lambda = 0, intercept = TRUE, scale = FALSE,
  init = NA, x_center = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{lambda}{numeric; not used}

\item{intercept}{logical; Whether an intercept should be included}

\item{scale}{logical; Should the features be normalized? (default: FALSE)}

\item{init}{numeric; not used}

\item{x_center}{logical;  Should the features be centered?}
}
\description{
Logistic Regression implementation that uses R's glm
}
