% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{logsumexp}
\alias{logsumexp}
\title{Numerically more stable way to calculate log sum exp}
\usage{
logsumexp(M)
}
\arguments{
\item{M}{matrix; m by n input matrix, sum with be over the rows}
}
\value{
matrix; m by 1 matrix
}
\description{
Numerically more stable way to calculate log sum exp
}
