% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Generics.R, R/NormalBasedClassifier.R
\docType{methods}
\name{losspart}
\alias{losspart}
\alias{losspart,NormalBasedClassifier-method}
\title{Loss of a classifier or regression function evaluated on partial labels}
\usage{
losspart(object, ...)

\S4method{losspart}{NormalBasedClassifier}(object, newdata, Y)
}
\arguments{
\item{object}{TODO}

\item{...}{additional parameters}

\item{newdata}{design matrix}

\item{Y}{class responsibility matrix}
}
\value{
numeric of the total loss on the test data
}
\description{
Loss of a classifier or regression function evaluated on partial labels
}

