% Generated by roxygen2 (4.0.2): do not edit by hand
\name{MajorityClassClassifier}
\alias{MajorityClassClassifier}
\title{Majority Class Classifier}
\usage{
MajorityClassClassifier(X, y, ...)
}
\arguments{
\item{X}{Design matrix, intercept term is added within the function}

\item{y}{Vector with class assignments}

\item{...}{Unused}
}
\description{
TODO
}

