% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ManifoldCorruptionLeastSquaresClassifier.R
\name{ManifoldCorruptionLeastSquaresClassifier}
\alias{ManifoldCorruptionLeastSquaresClassifier}
\title{Least Squares Classifier applying gaussian corruption over the manifold}
\usage{
ManifoldCorruptionLeastSquaresClassifier(X, y, X_u, lambda = 0,
  adjacency_kernel = vanilladot(), x_center = TRUE, scale = TRUE,
  y_scale = TRUE, k = 10, sigma = 1, samples = 100,
  method = "marginalize")
}
\description{
Least Squares Classifier applying gaussian corruption over the manifold
}

