% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErrorCurve.R
\name{measure_accuracy}
\alias{measure_accuracy}
\title{Generate Learning Curve for losses at different fractions of objects labeled}
\usage{
measure_accuracy(trained_classifier, X_l, y_l, X_u, y_u, X_test, y_test)
}
\arguments{
\item{X}{design matrix}

\item{y}{vector of labels}

\item{classifiers}{list; Classifiers to crossvalidate}

\item{fracs}{list; fractions of labeled data to use}

\item{with_replacement}{Sample objects with replacement}

\item{repeats}{integer; Number of learning curves to generate}

\item{verbose}{logical; Should progress to printed?}

\item{n_min}{integer; the minimum number of objects per class in the labeled set}

\item{test_fraction}{numeric; If not NULL a fraction of the object will be left out to serve as the test set}
}
\description{
Generate Learning Curve for losses at different fractions of objects labeled
}

