% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{missing_labels}
\alias{missing_labels}
\title{Access the true labels for the objects with missing labels when they are stored as an attribute in a data frame}
\usage{
missing_labels(df)
}
\arguments{
\item{df}{data.frame;}
}
\description{
Access the true labels for the objects with missing labels when they are stored as an attribute in a data frame
}

