\name{MMLeastSquaresClassifier}
\alias{MMLeastSquaresClassifier}
\title{LogisticLossClassifier}
\usage{
  MMLeastSquaresClassifier(X, y, X_u = NULL, lambda1 = 0,
    lambda2 = 0, intercept = TRUE, scale = FALSE,
    method = "default", y_scale = FALSE, lambda3 = 0,
    trueprob = NULL, ...)
}
\description{
  LogisticLossClassifier

  Minimax LSC
}

