% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/caretModels.R
\docType{data}
\name{model_svmlin}
\alias{model_svmlin}
\title{TODO: Document the usage of these functions}
\format{\preformatted{List of 13
 $ label     : chr "SVMlin"
 $ library   : chr "RSSL"
 $ type      : chr "Classification"
 $ parameters:'data.frame':	2 obs. of  3 variables:
  ..$ parameter: Factor w/ 2 levels "lambda","lambda_u": 1 2
  ..$ class    : Factor w/ 1 level "numeric": 1 1
  ..$ label    : Factor w/ 2 levels "Lambda","Lambda_u": 1 2
 $ grid      :function (x, y, len = NULL)  
 $ loop      : NULL
 $ fit       :function (x, y, wts, param, lev, last, classProbs, X_u, ...)  
 $ predict   :function (modelFit, newdata, submodels = NULL)  
 $ prob      :function (modelFit, newdata, preProc = NULL, submodels = NULL)  
 $ varImp    : NULL
 $ predictors:function (x, ...)  
 $ levels    :function (x)  
 $ sort      :function (x)  
}}
\usage{
model_svmlin
}
\description{
TODO: Document the usage of these functions
}
\keyword{datasets}

