\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,LeastSquaresClassifier,missing-method}
\alias{plot,SelfLearning,ANY-method}
\title{Plot function}
\usage{
plot(x, y, ...)

\S4method{plot}{LeastSquaresClassifier,missing}(x)

\S4method{plot}{SelfLearning,ANY}(x, X, y, X_u)
}
\arguments{
  \item{x}{LeastSquaresClassifier object}
}
\description{
Plot function

Plot method for LeastSquaresClassifier

Plot method for SelfLearning Classifier The plot function
plot the label assignments in each iteration
}

