% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{plot.CrossValidation}
\alias{plot.CrossValidation}
\title{Plot CrossValidation object}
\usage{
\method{plot}{CrossValidation}(x, y, ...)
}
\arguments{
\item{x}{CrossValidation object}

\item{y}{Not used}

\item{...}{Not used}
}
\description{
Plot CrossValidation object
}
