% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ErrorCurve.R
\name{plot.LearningCurve}
\alias{plot.LearningCurve}
\title{Plot LearningCurve object}
\usage{
plot.LearningCurve(x, legendsetting = "right", ...)
}
\arguments{
\item{x}{LearningCurve object OR a list of LearningCurve objects TODO: fix for list}

\item{legendsetting}{ggplot2 option Where the legend should be plotted (default: right)}

\item{...}{Unused}

\item{measurement}{The number of the measurement that should be plotted}

\item{dataset_names}{A vector with dataset names corresponding to the LearningCurve objects}

\item{classifier_names}{A vector with (usually shortened) classifier names}

\item{ncol}{integer; Number of columns in the plot}

\item{type}{integer; Type of graph (not currently used)}
}
\value{
a ggplot2 object containing the figure
}
\description{
Plot LearningCurve object generated by the LearningCurve or LearningCurveSSL functions
}
\details{
If multiple LearningCurve objects are supplied, the metadata from the first object will be used to draw the curves. Dataset names should be distinct
}

