% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Generics.R
\name{plot}
\alias{plot}
\title{Plot function}
\usage{
plot(x, y, ...)
}
\arguments{
\item{x}{Classifier or Regression object}

\item{y}{TODO}

\item{...}{Additional parameters}
}
\description{
Plot function
}

