% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ErrorCurve.R
\name{plotErrorCurve}
\alias{plotErrorCurve}
\title{Plot ErrorCurve object}
\usage{
plotErrorCurve(x, measurement = 1, legendsetting = "right",
  dataset_names = NULL, classifier_names = NULL, ncol = 2, ...)
}
\arguments{
\item{x}{ErrorCurve object OR a list of errorcurve objects TODO: fix for list}

\item{measurement}{The number of the measurement that should be plotted}

\item{legendsetting}{ggplot2 option Where the legend should be plotted (default: right)}

\item{dataset_names}{A vector with dataset names corresponding to the ErrorCurve objects}

\item{classifier_names}{A vector with (usually shortened) classifier names}

\item{ncol}{integer; Number of columns in the plot}

\item{...}{Unused}
}
\value{
a ggplot2 object constaining the figure
}
\description{
Plot ErrorCurve object generated by the ErrorCurve or ErrorCurveSSL functions
}
\details{
If multiple ErrorCurve objects are supplied, the metadata from the first object will be used to draw the curves. Dataset names should be distinct
}

