% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{posterior}
\alias{posterior}
\title{Class Posteriors of a classifier}
\usage{
posterior(object, ...)
}
\arguments{
\item{object}{Classifier or Regression object}

\item{...}{Additional parameters}
}
\description{
Class Posteriors of a classifier
}

