% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{print.CrossValidation}
\alias{print.CrossValidation}
\title{Print CrossValidation object}
\usage{
\method{print}{CrossValidation}(x, ...)
}
\arguments{
\item{x}{CrossValidation object}

\item{...}{Not used}
}
\description{
Print CrossValidation object
}
