% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningCurve.R
\name{print.LearningCurve}
\alias{print.LearningCurve}
\title{Print LearningCurve object}
\usage{
\method{print}{LearningCurve}(x, ...)
}
\arguments{
\item{x}{LearningCurve object}

\item{...}{Not used}
}
\description{
Print LearningCurve object
}
