% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Classifier.R, R/Generics.R, R/LeastSquaresClassifier.R, R/NormalBasedClassifier.R
\docType{methods}
\name{show,Classifier-method}
\alias{rssl-formatting}
\alias{show,Classifier-method}
\alias{show,LeastSquaresClassifier-method}
\alias{show,NormalBasedClassifier-method}
\title{Show the contents of a classifier}
\usage{
\S4method{show}{Classifier}(object)

\S4method{show}{LeastSquaresClassifier}(object)

\S4method{show}{NormalBasedClassifier}(object)
}
\arguments{
\item{object}{classifier}
}
\description{
Show the contents of a classifier

Show RSSL classifier
}

