% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSSL-package.R
\docType{package}
\name{RSSL-package}
\alias{RSSL}
\alias{RSSL-package}
\title{RSSL: Implementations of Semi-Supervised Learning Approaches for Classification}
\description{
A collection of implementations of semi-supervised classifiers and methods to evaluate their performance. The package includes implementations of, among others, Implicitly Constrained Learning, Moment Constrained Learning, the Transductive SVM, Manifold regularization, Maximum Contrastive Pessimistic Likelihood estimation, S4VM and WellSVM.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jkrijthe/RSSL}
  \item Report bugs at \url{https://github.com/jkrijthe/RSSL}
}

}
\author{
\strong{Maintainer}: Jesse Krijthe \email{jkrijthe@gmail.com}

}
\keyword{internal}
