% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GRFClassifier.R, R/Generics.R, R/KernelLeastSquaresClassifier.R, R/LeastSquaresClassifier.R, R/LinearSVM.R, R/LogisticLossClassifier.R, R/LogisticRegression.R, R/MajorityClassClassifier.R, R/NormalBasedClassifier.R, R/SVM.R, R/SVMlin.R, R/SelfLearning.R, R/TSVMcccp.R, R/TSVMuniversum.R, R/USMLeastSquaresClassifier.R
\docType{methods}
\name{predict,GRFClassifier-method}
\alias{predict,GRFClassifier-method}
\alias{predict,KernelLeastSquaresClassifier-method}
\alias{predict,LeastSquaresClassifier-method}
\alias{predict,LinearSVM-method}
\alias{predict,LogisticLossClassifier-method}
\alias{predict,LogisticRegression-method}
\alias{predict,MajorityClassClassifier-method}
\alias{predict,NormalBasedClassifier-method}
\alias{predict,SVM-method}
\alias{predict,SVMlin-method}
\alias{predict,SelfLearning-method}
\alias{predict,TSVMcccp-method}
\alias{predict,TSVMuniversvm-method}
\alias{predict,USMLeastSquaresClassifier-method}
\alias{rssl-predict}
\title{Predict using RSSL classifier}
\usage{
\S4method{predict}{GRFClassifier}(object, newdata, ...)

\S4method{predict}{KernelLeastSquaresClassifier}(object, newdata,
  probs = FALSE, ...)

\S4method{predict}{LeastSquaresClassifier}(object, newdata, probs = FALSE,
  ...)

\S4method{predict}{LinearSVM}(object, newdata)

\S4method{predict}{LogisticLossClassifier}(object, newdata, probs = FALSE)

\S4method{predict}{LogisticRegression}(object, newdata, probs = FALSE)

\S4method{predict}{MajorityClassClassifier}(object, newdata, probs = FALSE)

\S4method{predict}{NormalBasedClassifier}(object, newdata, probs = FALSE)

\S4method{predict}{SVM}(object, newdata)

\S4method{predict}{SVMlin}(object, newdata, probs = FALSE)

\S4method{predict}{SelfLearning}(object, newdata, ...)

\S4method{predict}{TSVMcccp}(object, newdata, probs = FALSE)

\S4method{predict}{TSVMuniversvm}(object, newdata, probs = FALSE)

\S4method{predict}{USMLeastSquaresClassifier}(object, newdata, probs = FALSE,
  ...)
}
\arguments{
\item{object}{classifier}

\item{newdata}{objects to generate predictions for}

\item{...}{Other arguments}

\item{probs}{Whether probabilities should be returned}
}
\description{
Predict using RSSL classifier

Log likelihood on the training set

For the SelfLearning Classifier the Predict Method delegates prediction to the specific model object
}

