% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningCurve.R
\name{sample_k_per_level}
\alias{sample_k_per_level}
\title{Sample k indices per levels from a factor}
\usage{
sample_k_per_level(y, k)
}
\arguments{
\item{y}{factor; factor with levels}

\item{k}{integer; number of indices to sample per level}
}
\value{
vector with indices for sample
}
\description{
Sample k indices per levels from a factor
}
