% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleMatrix.R
\name{scaleMatrix}
\alias{scaleMatrix}
\title{matrix scaling inspired by stdize from the PLS package}
\usage{
scaleMatrix(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{matrix to be standardized}

\item{center}{TRUE if x should be centered}

\item{scale}{logical; TRUE of x should be scaled by the standard deviation}
}
\description{
matrix scaling inspired by stdize from the PLS package
}

