\name{show}
\alias{show}
\alias{show,Classifier-method}
\alias{show,NormalBasedClassifier-method}
\title{Show method for Classifier}
\arguments{
  \item{object}{object of class NormalBasedClassifier}
}
\description{
  Show method for Classifier

  Print some properties of classifiers based on classes
  modeled as Gaussians.
}

