% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/TSVMCCPte.R, R/TSVM_CCCP.R
\name{solve_svm}
\alias{solve_svm}
\title{SVM solve.QP implementation}
\usage{
solve_svm(K, y, C = 1)

solve_svm(K, y, C = 1)
}
\description{
SVM solve.QP implementation

SVM solve.QP implementation
}

