% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SSLDataFrameToMatrices}
\alias{SSLDataFrameToMatrices}
\title{Convert data.frame to matrices for semi-supervised learners

Given a formula object and a data.frame, extract the design matrix X for the labeled observations, X_u for the unlabeled observations and y for the labels of the labeled observations}
\usage{
SSLDataFrameToMatrices(model,D,intercept=TRUE)
}
\arguments{
\item{model}{Formula object with model}

\item{D}{data.frame object with objects}

\item{intercept}{Whether to include an intercept in the design matrices}
}
\value{
list object with the following objects:
\item{X}{design matrix of the labeled data}
\item{X_u}{design matrix of the unlabeled data}
\item{y}{integer vector indicating the labels of the labeled data}
\item{classnames}{names of the classes corresponding to the integers in y}
}
\description{
Convert data.frame to matrices for semi-supervised learners

Given a formula object and a data.frame, extract the design matrix X for the labeled observations, X_u for the unlabeled observations and y for the labels of the labeled observations
}

