% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{stderror}
\alias{stderror}
\title{Standard Error of a vector}
\usage{
stderror(x)
}
\arguments{
\item{x}{numeric; vector for which to calculate standard error}
}
\description{
Standard Error of a vector
}

