% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{summary.CrossValidation}
\alias{summary.CrossValidation}
\title{Summary of Crossvalidation results}
\usage{
\method{summary}{CrossValidation}(object, measure = NULL, ...)
}
\arguments{
\item{object}{CrossValidation object}

\item{measure}{Measure of interest}

\item{...}{Not used}
}
\description{
Summary of Crossvalidation results
}
