% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPLDA.R
\name{svdinv}
\alias{svdinv}
\title{Inverse of a matrix using the singular value decomposition}
\usage{
svdinv(X)
}
\arguments{
\item{X}{matrix; square input matrix}
}
\value{
Y matrix; inverse of the input matrix
}
\description{
Inverse of a matrix using the singular value decomposition
}
