% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPLDA.R
\name{svdinvsqrtm}
\alias{svdinvsqrtm}
\title{Taking the inverse of the square root of the matrix using the singular value decomposition}
\usage{
svdinvsqrtm(X)
}
\arguments{
\item{X}{matrix; square input matrix}
}
\value{
Y matrix; inverse of the square root of the input matrix
}
\description{
Taking the inverse of the square root of the matrix using the singular value decomposition
}
