% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPLDA.R
\name{svdsqrtm}
\alias{svdsqrtm}
\title{Taking the square root of a matrix using the singular value decomposition}
\usage{
svdsqrtm(X)
}
\arguments{
\item{X}{matrix; square input matrix}
}
\value{
Y matrix; square root of the input matrix
}
\description{
Taking the square root of a matrix using the singular value decomposition
}
