% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SVM.R
\name{SVM}
\alias{SVM}
\title{SVM Classifier}
\usage{
SVM(X, y, C = 1, method = "Dual", scale = TRUE, intercept = TRUE,
  kernel = NULL, ...)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Labels of the observations}

\item{C}{Cost variable}

\item{method}{Estimation procedure c("Dual","Primal","BGD","SGD","Pegasos")}

\item{scale}{Whether a z-transform should be applied (default: TRUE)}

\item{intercept}{Whether an intercept should be added (default: TRUE)}
}
\value{
S4 object of type LinearSVM
}
\description{
SVM Classifier
}

