% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svmlin.R
\name{svmlin}
\alias{svmlin}
\title{svmlin implementation by Sindhwani}
\usage{
svmlin(X, y, Xu, algorithm = 1, lambda = 1, lambda_u = 1,
  max_switch = 10000, pos_frac = 0.5, Cp = 1, Cn = 1, verbose = FALSE)
}
\arguments{
\item{X}{Matrix or sparseMatrix containing the labeled feature vectors, without intercept}

\item{y}{factor containing class assignments}

\item{Xu}{Matrix or sparseMatrix containing the unlabeled feature vectors, without intercept}

\item{lambda}{regularization parameter lambda (default 1)}

\item{lambda_u}{regularization parameter lambda_u (default 1)}

\item{max_switch}{-S maximum number of switches in TSVM (default 10000)}

\item{pos_frac}{positive class fraction of unlabeled data  (default 0.5)}

\item{Cp}{relative cost for positive examples (only available with algorithm 1)}

\item{Cn}{relative cost for positive examples (only available with algorithm 1)}

\item{Algorithm}{choice, see details (default:1)}
}
\description{
R interface to the svmlin software by Vikas Sindhwani for fast linear transductive SVMs.
}
\details{
Algorithms
0 -- Regularized Least Squares Classification (RLSC)
1 -- SVM (L2-SVM-MFN) (default choice)
2 -- Multi-switch Transductive SVM (using L2-SVM-MFN)
3 -- Deterministic Annealing Semi-supervised SVM (using L2-SVM-MFN)
}
\examples{
data(svmlin_example)
t_svmlin_1 <- svmlin(svmlin_example$X_train[1:50,],svmlin_example$y_train,Xu=NULL, lambda = 0.001)
t_svmlin_2 <- svmlin(svmlin_example$X_train[1:50,],svmlin_example$y_train,Xu=svmlin_example$X_train[-c(1:50),], lambda = 10,lambda_u=100,algorithm = 2)
# Calculate Accuracy
mean(predict(t_svmlin_1,svmlin_example$X_test)==svmlin_example$y_test)
mean(predict(t_svmlin_2,svmlin_example$X_test)==svmlin_example$y_test)
}
\references{
Vikas Sindhwani and S. Sathiya Keerthi. Large Scale Semi-supervised Linear SVMs.     Proceedings of ACM SIGIR, 2006
 @references V. Sindhwani and S. Sathiya Keerthi. Newton Methods for Fast Solution of Semi-supervised Linear SVMs. Book Chapter in Large Scale Kernel Machines, MIT Press, 2006
}

