% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testdata-data.R
\docType{data}
\name{testdata}
\alias{testdata}
\title{Labeled training design matrix}
\description{
A list containing a sample from the \code{GenerateSlicedCookie} dataset for unit testing and examples.
}

