% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WellSVM.R
\name{threshold}
\alias{threshold}
\title{Refine the prediction to satisfy the balance constraint}
\usage{
threshold(y1, options)
}
\arguments{
\item{y1}{predictions}

\item{options}{options passed}
}
\value{
y2
}
\description{
Refine the prediction to satisfy the balance constraint
}
