% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{true_labels}
\alias{true_labels}
\title{Access the true labels when they are stored as an attribute in a data frame}
\usage{
true_labels(df)
}
\arguments{
\item{df}{data.frame;}
}
\description{
Access the true labels when they are stored as an attribute in a data frame
}

