% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSVM.R
\name{TSVM}
\alias{TSVM}
\title{Transductive SVM classifier using the convex concave procedure}
\usage{
TSVM(X, y, X_u, C, Cstar, kernel = kernlab::vanilladot(),
  balancing_constraint = TRUE, s = 0, x_center = TRUE, scale = FALSE,
  eps = 1e-09, verbose = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{C}{numeric; Cost parameter of the SVM}

\item{Cstar}{numeric; Cost parameter of the unlabeled objects}

\item{kernel}{kernlab::kernel to use}

\item{balancing_constraint}{logical; Whether a balancing constraint should be enfored that causes the fraction of objects assigned to each label in the unlabeled data to be similar to the label fraction in the labeled data.}

\item{s}{numeric; parameter controlling the loss function of the unlabeled objects (generally values between -1 and 0)}

\item{x_center}{logical;  Should the features be centered?}

\item{scale}{If TRUE, apply a z-transform to all observations in X and X_u before running the regression}

\item{eps}{numeric; Stopping criterion for the maximinimization}

\item{verbose}{logical; print debugging messages, only works for vanilladot() kernel (default: FALSE)}
}
\description{
Transductive SVM using the CCCP algorithm as proposed by Collobert et al. (2006) implemented in R using the quadprog package. The implementation does not handle large datasets very well, but can be useful for visualization purposes.
}
\references{
Collobert, R. et al., 2006. Large scale transductive SVMs. Journal of Machine Learning Research, 7, pp.1687-1712.
}

