% Generated by roxygen2 (4.0.2): do not edit by hand
\name{TSVM_CCCPlin}
\alias{TSVM_CCCPlin}
\title{CCCP Transductive SVM classifier}
\usage{
TSVM_CCCPlin(X, y, X_u, C, Cstar, s = -0.3, x_center = FALSE,
  scale = FALSE, eps = 1e-06, ...)
}
\arguments{
\item{X}{matrix; Design matrix, intercept term is added within the function}

\item{y}{vector; Vector or factor with class assignments}

\item{X_u}{matrix; Design matrix of the unlabeled data, intercept term is added within the function}

\item{scale}{If TRUE, apply a z-transform to all observations in X and X_u before running the regression}

\item{...}{additional arguments}

\item{Clabeled}{numeric; Cost parameter of the SVM}

\item{Cunlabeled}{numeric; Cost parameter of the unlabeled objects}

\item{fracpos}{numeric; fraction of test points to assign to the first class}

\item{intercept}{TRUE if an intercept should be added to the model}
}
\value{
S4 object of class TSVM with the following slots:
\item{theta}{weight vector}
\item{classnames}{the names of the classes}
\item{modelform}{formula object of the model used in regression}
\item{scaling}{a scaling object containing the paramters of the z-transforms applied to the data}
\item{optimization}{the object returned by the optim function}
\item{unlabels}{the labels assigned to the unlabeled objects}
}
\description{
CCCP Transductive SVM classifier
}

