% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSVMcccp.R
\name{TSVMcccp}
\alias{TSVMcccp}
\title{CCCP Transductive SVM classifier}
\usage{
TSVMcccp(X, y, X_u, C, Cstar, s = -0.3, x_center = FALSE, scale = FALSE,
  eps = 1e-09, verbose = FALSE, ...)
}
\arguments{
\item{X}{matrix; Design matrix, intercept term is added within the function}

\item{y}{vector; Vector or factor with class assignments}

\item{X_u}{matrix; Design matrix of the unlabeled data, intercept term is added within the function}

\item{C}{numeric; Cost parameter of the SVM}

\item{Cstar}{numeric; Cost parameter of the unlabeled objects}

\item{s}{numeric; parameter controlling the loss function of the unlabeled objects}

\item{x_center}{logical;  Should the features be centered?}

\item{scale}{If TRUE, apply a z-transform to all observations in X and X_u before running the regression}

\item{eps}{numeric; Stopping criterion for the maximinimization}

\item{verbose}{logical; print debugging messages (default: FALSE)}

\item{...}{additional arguments}
}
\value{
S4 object of class TSVM with the following slots:
\item{theta}{weight vector}
\item{classnames}{the names of the classes}
\item{modelform}{formula object of the model used in regression}
\item{scaling}{a scaling object containing the paramters of the z-transforms applied to the data}
\item{optimization}{the object returned by the optim function}
\item{unlabels}{the labels assigned to the unlabeled objects}
}

