% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSVMuniversum.R
\name{TSVMuniversvm}
\alias{TSVMuniversvm}
\title{TSVM universvm implementation}
\usage{
TSVMuniversvm(X, y, X_u = NULL, x_center = FALSE, scale = FALSE,
  binary_path = NULL, temp_path = NULL, ...)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{x_center}{logical;  Should the features be centered?}

\item{scale}{boolean; Should the features be normalized? (default: FALSE)}

\item{binary_path}{character; Path to universvm binary}

\item{temp_path}{character; Path to temp directory}

\item{...}{Not used}
}
\description{
TSVM universvm implementation
}
\seealso{
Other RSSL classifiers: \code{\link{MCLinearDiscriminantClassifier}};
  \code{\link{MCNearestMeanClassifier}}
}

