% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WellSVM.R
\name{WellSVM_SSL}
\alias{WellSVM_SSL}
\title{Convex relaxation of S3VM by label generation}
\usage{
WellSVM_SSL(K0, y, opt, yinit = NULL)
}
\arguments{
\item{K0}{kernel matrix}

\item{y}{labels}

\item{opt}{options}

\item{yinit}{label initialization (not used)}
}
\description{
Convex relaxation of S3VM by label generation
}
