% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WellSVM.R
\name{WellSVM_supervised}
\alias{WellSVM_supervised}
\title{A degenerated version of WellSVM where the labels are complete, that is, supervised learning}
\usage{
WellSVM_supervised(K0, y, opt, ind_y)
}
\arguments{
\item{K0}{kernel matrix}

\item{y}{labels}

\item{opt}{options}

\item{ind_y}{Labeled/Unlabeled indicator}
}
\description{
A degenerated version of WellSVM where the labels are complete, that is, supervised learning
}
