% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPLDA.R
\name{wlda_loglik}
\alias{wlda_loglik}
\title{Measures the expected log-likelihood of the LDA model defined by m, p,
and iW on the data set a, where weights w are potentially taken into
account}
\usage{
wlda_loglik(m, p, iW, a, w)
}
\arguments{
\item{m}{means}

\item{p}{class prior}

\item{iW}{is the inverse of the within covariance matrix}

\item{a}{design matrix}

\item{w}{weights}
}
\value{
Average log likelihood
}
\description{
Measures the expected log-likelihood of the LDA model defined by m, p,
and iW on the data set a, where weights w are potentially taken into
account
}
