% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{xtable.CrossValidation}
\alias{xtable.CrossValidation}
\title{Generate Latex table of cross-validation results}
\usage{
\method{xtable}{CrossValidation}(object, caption = "",
  benchmark_method = NULL, exclude_methods = NULL)
}
\arguments{
\item{object}{cross-validation results object}

\item{caption}{character; Caption to be placed over the latex table}

\item{benchmark_method}{character; The method to be used to compare others with}

\item{exclude_methods}{character vector; Methods to exclude from the table}
}
\description{
Generate Latex table of cross-validation results
}

