\docType{methods}
\name{boundaryplot}
\alias{boundaryplot}
\alias{boundaryplot,LeastSquaresClassifier-method}
\alias{boundaryplot,LinearSVM-method}
\alias{boundaryplot,LogisticLossClassifier-method}
\alias{boundaryplot,LogisticRegression-method}
\alias{boundaryplot,NormalBasedClassifier-method}
\alias{boundaryplot,SGDSVM-method}
\title{Decision Boundary plot of a classifier}
\usage{
boundaryplot(object, ...)

\S4method{boundaryplot}{LeastSquaresClassifier}(object, p)

\S4method{boundaryplot}{LogisticLossClassifier}(object, p)

\S4method{boundaryplot}{LogisticRegression}(object, p)

\S4method{boundaryplot}{NormalBasedClassifier}(object, p, ...)

\S4method{boundaryplot}{SGDSVM}(object, p)

\S4method{boundaryplot}{LinearSVM}(object, p)
}
\arguments{
  \item{object}{LeastSquaresClassifier object}

  \item{p}{ggplot object of classification problem
  generated by clplot}

  \item{object}{LogisticLossClassifier object}

  \item{p}{ggplot object of classification problem
  generated by clplot}

  \item{object}{LeastSquaresClassifier object}

  \item{p}{ggplot object of classification problem
  generated by clplot}

  \item{object}{SGDSVM object}

  \item{p}{ggplot object of classification problem
  generated by clplot}
}
\description{
Decision Boundary plot of a classifier

Boundary plot method for LeastSquaresClassifier

Boundary plot method for LogisticLossClassifier

Boundary plot method for LogisticRegression

Boundary plot method for LeastSquaresClassifier

Boundary plot method for SGDSVM

Predict class of new observations using a LinearSVM
}

