% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/LogisticRegression.R
\docType{methods}
\name{boundaryplot}
\alias{boundaryplot}
\alias{boundaryplot,LogisticRegression-method}
\title{Decision Boundary plot of a classifier}
\usage{
boundaryplot(object, ...)

\S4method{boundaryplot}{LogisticRegression}(object, p)
}
\arguments{
\item{object}{object to plot}

\item{...}{Additional options}

\item{p}{ggplot object of classification problem generated by clplot}
}
\description{
Decision Boundary plot of a classifier
}

