% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Generics.R, R/LeastSquaresClassifier.R, R/LinearSVM.R, R/LogisticLossClassifier.R, R/LogisticRegression.R, R/NormalBasedClassifier.R, R/SVM.R
\docType{methods}
\name{boundaryplot}
\alias{boundaryplot}
\alias{boundaryplot,LeastSquaresClassifier-method}
\alias{boundaryplot,LinearSVM-method}
\alias{boundaryplot,LogisticLossClassifier-method}
\alias{boundaryplot,LogisticRegression-method}
\alias{boundaryplot,NormalBasedClassifier-method}
\title{Decision Boundary plot of a classifier}
\usage{
boundaryplot(object, ...)

\S4method{boundaryplot}{LeastSquaresClassifier}(object, p)

\S4method{boundaryplot}{LinearSVM}(object, p)

\S4method{boundaryplot}{LogisticLossClassifier}(object, p)

\S4method{boundaryplot}{LogisticRegression}(object, p)

\S4method{boundaryplot}{NormalBasedClassifier}(object, p, ...)

\S4method{boundaryplot}{LinearSVM}(object, p)
}
\arguments{
\item{object}{object to plot}

\item{...}{Additional options}

\item{p}{ggplot object of classification problem generated by clplot}
}
\description{
Decision Boundary plot of a classifier
}

