% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{cov_ml}
\alias{cov_ml}
\title{Biased (maximum likelihood) estimate of the covariance matrix}
\usage{
cov_ml(X)
}
\arguments{
\item{X}{matrix with observations}
}
\description{
Biased (maximum likelihood) estimate of the covariance matrix
}
