\name{CrossValidation}
\alias{CrossValidation}
\title{Regular cross-validation on a set of classifiers}
\usage{
CrossValidation(X, y, classifiers, measures = c("predict", "losstest"),
  groups = NULL, k = 10, repeats = 1, verbose = FALSE)
}
\description{
Regular cross-validation on a set of classifiers
}

