% Generated by roxygen2 (4.0.2): do not edit by hand
\name{CrossValidationTransductive}
\alias{CrossValidationTransductive}
\title{Cross-validation in the transductive setting}
\usage{
CrossValidationTransductive(X, y, classifiers, measures = c("predict",
  "losstest"), groups = NULL, k = 10, repeats = 1, verbose = FALSE,
  dataset_name = "Unknown Dataset")
}
\arguments{
\item{X}{design matrix of the labeled objects}

\item{y}{vector with labels}

\item{classifiers}{list; Classifiers to crossvalidate}

\item{measures}{character; c("predict","losstest") giving the measures to be used}

\item{groups}{vector; Group assigments to be taken into account in defining the folds}

\item{k}{integer; Number of folds in the cross-validation}

\item{repeats}{integer; Number of repeated assignments to folds}

\item{verbose}{logical; Controls the verbosity of the output}

\item{dataset_name}{character; Name of the dataset}
}
\description{
Cross-validation in the transductive setting
}

