% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/LeastSquaresClassifier.R, R/KernelLeastSquaresClassifier.R, R/LinearSVM.R, R/SVM.R
\docType{methods}
\name{decisionvalues}
\alias{decisionvalues}
\alias{decisionvalues,KernelLeastSquaresClassifier-method}
\alias{decisionvalues,LeastSquaresClassifier-method}
\alias{decisionvalues,LinearSVM-method}
\alias{decisionvalues,SVM-method}
\title{Decision values returned by a classifier for a set of objects}
\usage{
decisionvalues(object, newdata)

\S4method{decisionvalues}{LeastSquaresClassifier}(object, newdata)

\S4method{decisionvalues}{KernelLeastSquaresClassifier}(object, newdata)

\S4method{decisionvalues}{LinearSVM}(object, newdata)

\S4method{decisionvalues}{SVM}(object, newdata)
}
\arguments{
\item{object}{Classifier object}

\item{newdata}{new data to classify}
}
\description{
Returns decision values of a classifier
}

