% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{df_to_matrices}
\alias{df_to_matrices}
\title{Convert data.frame with missing labels to matrices}
\usage{
df_to_matrices(df, formula = NULL)
}
\arguments{
\item{df}{data.frame; Data}

\item{formula}{formula; Description of problem}
}
\description{
Convert data.frame with missing labels to matrices
}
\seealso{
Other RSSL utilities: 
\code{\link{LearningCurveSSL}()},
\code{\link{SSLDataFrameToMatrices}()},
\code{\link{add_missinglabels_mar}()},
\code{\link{measure_accuracy}()},
\code{\link{missing_labels}()},
\code{\link{split_dataset_ssl}()},
\code{\link{split_random}()},
\code{\link{true_labels}()}
}
\concept{RSSL utilities}
