% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMLeastSquaresClassifier.R
\name{EMLeastSquaresClassifier}
\alias{EMLeastSquaresClassifier}
\title{An Expectation Maximization like approach to Semi-Supervised Least Squares Classification}
\usage{
EMLeastSquaresClassifier(X, y, X_u, x_center = FALSE, scale = FALSE,
  verbose = FALSE, intercept = TRUE, lambda = 0, eps = 1e-09,
  y_scale = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{x_center}{logical;  Should the features be centered?}

\item{scale}{Should the features be normalized? (default: FALSE)}

\item{verbose}{logical; Controls the verbosity of the output}

\item{intercept}{logical; Whether an intercept should be included}

\item{lambda}{numeric; L2 regularization parameter}

\item{eps}{Stopping criterion for the minimization}

\item{y_scale}{logical; whether the target vector should be centered}

\item{...}{Additional Parameters, Not used}
}
\description{
Minimize the total loss of the labeled and unlabeled objects by finding the weight vector and labels that minimize the total loss. The algorithm proceeds similar to EM, by subsequently applying a weight update and a soft labeling of the unlabeled objects. This is repeated until convergence.
}
\seealso{
Other RSSL LeastSquares: \code{\link{ICLeastSquaresClassifier}}
}

