% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMLinearDiscriminantClassifier.R
\name{EMLinearDiscriminantClassifier}
\alias{EMLinearDiscriminantClassifier}
\title{Expectation Maximization Nearest Mean Classifier}
\usage{
EMLinearDiscriminantClassifier(X, y, X_u, method = "EM", scale = FALSE,
  eps = 1e-08, verbose = FALSE, ...)
}
\arguments{
\item{X}{design matrix of the labeled objects}

\item{y}{vector with labels}

\item{X_u}{design matrix of the labeled objects}

\item{method}{character; Currently only "EM"}

\item{scale}{Should the features be normalized? (default: FALSE)}

\item{eps}{Stopping criterion for the maximinimization}

\item{verbose}{logical; Controls the verbosity of the output}

\item{...}{Additional Parameters, Not used}
}
\description{
Expectation Maximization Nearest Mean Classifier
}

