% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EMLinearDiscriminantClassifier.R
\name{EMLinearDiscriminantClassifier}
\alias{EMLinearDiscriminantClassifier}
\title{Semi-Supervised Linear Discriminant Analysis using Expectation Maximization}
\usage{
EMLinearDiscriminantClassifier(X, y, X_u, method = "EM", scale = FALSE,
  eps = 1e-08, verbose = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{method}{character; Currently only "EM"}

\item{scale}{boolean; Should the features be normalized? (default: FALSE)}

\item{eps}{Stopping criterion for the maximinimization}

\item{verbose}{logical; Controls the verbosity of the output}
}
\description{
Expectation Maximization
}

