% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/EMNearestMeanClassifier.R
\name{EMNearestMeanClassifier}
\alias{EMNearestMeanClassifier}
\title{Expectation Maximization Nearest Mean Classifier}
\usage{
EMNearestMeanClassifier(X, y, X_u, method = "EM", scale = FALSE,
  eps = 1e-04, ...)
}
\arguments{
\item{X}{design matrix of the labeled objects}

\item{y}{vector with labels}

\item{X_u}{design matrix of the labeled objects}

\item{method}{character; Currently only "EM"}

\item{scale}{Should the features be normalized? (default: FALSE)}

\item{eps}{Stopping criterion for the maximinimization}

\item{...}{Unused}
}
\description{
Expectation Maximization Nearest Mean Classifier
}

