% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ErrorCurve.R
\name{ErrorCurveSSL}
\alias{ErrorCurveSSL}
\title{Generate Semi-Supervised Learning error curve}
\usage{
ErrorCurveSSL(X, y, classifiers, n_l, with_replacement = FALSE,
  sizes = 2^(1:8), n_test = 1000, repeats = 100, verbose = FALSE,
  n_min = 1, dataset_name = NULL)
}
\arguments{
\item{X}{design matrix}

\item{y}{vector of labels}

\item{classifiers}{list; Classifiers to crossvalidate}

\item{n_l}{Number of labeled objects to be used in the experiments}

\item{with_replacement}{Indicated whether the subsampling is done with replacement or not (default: FALSE)}

\item{sizes}{vector with number of unlabeled objects for which to evaluate performance}

\item{n_test}{Number of test points if with_replacement is TRUE}

\item{repeats}{Number of learning curves to draw}

\item{verbose}{Print progressbar during execution (default: FALSE)}

\item{n_min}{Minimum number of labeled objects per class in}

\item{dataset_name}{character; Name of the dataset}
}
\value{
ErrorCurve object
}
\description{
Generate Semi-Supervised Learning error curve
}

