\name{GenerateFourClusters}
\alias{GenerateFourClusters}
\title{Generate Four Clusters dataset}
\usage{
GenerateFourClusters(n=100,distance=6,expected=FALSE)
}
\arguments{
  \item{n}{number of observations to generate}

  \item{distance}{Distance between clusters (default: 6)}

  \item{expected}{TRUE if the large margin equals the class
  boundary, FALSE if the class boundary is perpendicular to
  the large margin}
}
\description{
Generate a four clusters dataset
}

