% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GenerateSSLData.R
\name{generateTwoCircles}
\alias{generateTwoCircles}
\title{Generate data from 2 circles}
\usage{
generateTwoCircles(n = 100, noise_var = 0.2)
}
\arguments{
\item{n}{integer; Number of examples to generate}

\item{noise_var}{numeric; size of the variance parameter}

\item{expected}{logical; whether the decision boundary should be the expected or perpendicular}
}
\description{
One circle circumscribes the other
}

