% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateSSLData.R
\name{generateTwoCircles}
\alias{generateTwoCircles}
\title{Generate data from 2 circles}
\usage{
generateTwoCircles(n = 100, noise_var = 0.2)
}
\arguments{
\item{n}{integer; Number of examples to generate}

\item{noise_var}{numeric; size of the variance parameter}
}
\description{
One circle circumscribes the other
}
\seealso{
Other RSSL datasets: 
\code{\link{generate2ClassGaussian}()},
\code{\link{generateABA}()},
\code{\link{generateCrescentMoon}()},
\code{\link{generateFourClusters}()},
\code{\link{generateParallelPlanes}()},
\code{\link{generateSlicedCookie}()},
\code{\link{generateSpirals}()}
}
\concept{RSSL datasets}
