% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plotting.R
\name{geom_classifier}
\alias{geom_classifier}
\title{Plot RSSL classifier boundary (deprecated)}
\usage{
geom_classifier(..., show_guide = TRUE)
}
\arguments{
\item{...}{List of trained classifiers}

\item{show_guide}{logical (default: TRUE); Show legend}
}
\description{
Deprecated: Use geom_linearclassifier or stat_classifier to plot classification boundaries
}
