% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/GRFClassifier.R
\name{GRFClassifier}
\alias{GRFClassifier}
\title{Gaussian Random Fields and Harmonic functions}
\usage{
GRFClassifier(X, y, X_u, sigma = 0.1, eta = 0.1, CMN = TRUE,
  scale = FALSE, x_center = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{eta}{The influence of the external classifier}

\item{CMN}{Should the Class Mass Normalization heuristic be applied? (default: TRUE)}

\item{kernel}{TODO: kernel function??}

\item{y_u}{Labels given by external classifier}
}
\description{
Gaussian Random Fields and Harmonic functions
}

