\name{ICLDA}
\alias{ICLDA}
\title{Moment constriant semi-supervised linear discriminant analysis}
\usage{
  MCLinearDiscriminantClassifier(X, y, X_u,
    method="closedform",prior=NULL, scale=FALSE, ...)
}
\arguments{
  \item{X}{<what param does>}

  \item{y}{<what param does>}

  \item{X_u}{<what param does>}

  \item{method}{<what param does>}

  \item{prior}{<what param does>}

  \item{scale}{<what param does>}

  \item{...}{<what param does>}
}
\description{
  <full description>
}

