\name{ImputedLeastSquaresClassifier}
\alias{ImputedLeastSquaresClassifier}
\title{Simple Imputed Least Squares Classifier}
\usage{
ImputedLeastSquaresClassifier(X, y, X_u, method = "average")
}
\arguments{
  \item{X}{Design matrix}

  \item{y}{labels}

  \item{X_u}{Design matrix unlabeled data}
}
\description{
Simple Imputed Least Squares Classifier
}

