% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ImputedLeastSquaresClassifier.R
\name{ImputedLeastSquaresClassifier}
\alias{ImputedLeastSquaresClassifier}
\title{Simple Imputed Least Squares Classifier}
\usage{
ImputedLeastSquaresClassifier(X, y, X_u, method = "average")
}
\arguments{
\item{X}{Design matrix}

\item{y}{labels}

\item{X_u}{Design matrix unlabeled data}

\item{method}{TODO}
}
\description{
Simple Imputed Least Squares Classifier
}

