\name{IWLogisticRegression}
\alias{IWLogisticRegression}
\title{Importance Weighted Logistic Regression}
\usage{
  IWLogisticRegression(X, y, X_u)
}
\arguments{
  \item{X}{Design matrix}

  \item{y}{labels}

  \item{X_u}{Design matrix unlabeled data}
}
\description{
  Implementation of the Importance Weighted Logistic
  Regression idea presented in Sokolovska2008
}

