% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LaplacianSVM.R
\name{LaplacianSVM}
\alias{LaplacianSVM}
\title{SVM Classifier}
\usage{
LaplacianSVM(X, y, X_u = NULL, lambda = 1, gamma = 1, scale = TRUE,
  kernel = vanilladot(), adjacency_kernel = rbfdot(1/4))
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{lambda}{numeric; L2 regularization parameter}

\item{gamma}{numeric; Weight of the unlabeled data}

\item{scale}{boolean; Should the features be normalized? (default: FALSE)}

\item{kernel}{kernlab::kernel to use}

\item{adjacency_kernel}{kernlab::kernel to use as adjacency kernel}
}
\value{
S4 object of type LaplacianSVM
}
\description{
SVM Classifier
}

