% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningCurve.R
\name{LearningCurveSSL}
\alias{LearningCurveSSL}
\alias{LearningCurveSSL.matrix}
\title{Generate Semi-Supervised learning curve.}
\usage{
LearningCurveSSL(X, y, ...)

\method{LearningCurveSSL}{matrix}(X, y, classifiers, measures = list(Accuracy
  = measure_accuracy), type = "unlabeled", n_l, with_replacement = FALSE,
  sizes = 2^(1:8), n_test = 1000, repeats = 100, verbose = FALSE,
  n_min = 1, dataset_name = NULL, test_fraction = NULL, fracs = seq(0.1,
  0.9, 0.1), time = TRUE, pre_scale = FALSE, pre_pca = FALSE,
  low_level_cores = 1, ...)
}
\arguments{
\item{X}{design matrix}

\item{y}{vector of labels}

\item{...}{arguments passed to underlying function}

\item{classifiers}{list; Classifiers to crossvalidate}

\item{measures}{named list of functions giving the measures to be used}

\item{type}{Type of learning curve, either "unlabeled" or "fraction"}

\item{n_l}{Number of labeled objects to be used in the experiments (see details)}

\item{with_replacement}{Indicated whether the subsampling is done with replacement or not (default: FALSE)}

\item{sizes}{vector with number of unlabeled objects for which to evaluate performance}

\item{n_test}{Number of test points if with_replacement is TRUE}

\item{repeats}{Number of learning curves to draw}

\item{verbose}{Print progressbar during execution (default: FALSE)}

\item{n_min}{Minimum number of labeled objects per class in}

\item{dataset_name}{character; Name of the dataset}

\item{test_fraction}{numeric; If not NULL a fraction of the object will be left out to serve as the test set}

\item{fracs}{list; fractions of labeled data to use}

\item{time}{logical; Whether execution time should be saved.}

\item{pre_scale}{logical; Whether the features should be scaled before the dataset is used}

\item{pre_pca}{logical; Whether the features should be preprocessed using a PCA step}
}
\value{
LearningCurve object
}
\description{
Evaluate semi-supervised classifiers for different amounts of unlabeled training examples or different fractions of unlabeled vs. labeled examples.
}
\details{
For n_l, additional options include: "enough" which takes the max of the number of features and 20, max(ncol(X)+5,20), "d" which takes the number of features or "2d" which takes 2 times the number of features.
}
\examples{
X <- model.matrix(Species~.-1,data=iris)
y <- iris$Species

classifiers <- list("LS"=function(X,y,X_u,y_u) {LeastSquaresClassifier(X,y,lambda=0)}, 
                    "RLS"=function(X,y,X_u,y_u) {LeastSquaresClassifier(X,y,lambda=10)})
measures <- list("Accuracy" =  measure_accuracy,
                 "Loss" = measure_losstest,
                 "Loss labeled" = measure_losslab,
                 "Loss Lab+Unlab" = measure_losstrain
                 )
lc <- LearningCurveSSL(X,y,classifiers=classifiers,measures=measures,n_l=10,repeats=5)
print(lc)
plot(lc)

}

