\name{LeastSquaresRegression}
\alias{LeastSquaresRegression}
\title{<brief desc>}
\usage{
  LeastSquaresRegression(X, y, lambda = 0, scale = FALSE,
    ...)
}
\arguments{
  \item{X}{<what param does>}

  \item{y}{<what param does>}

  \item{lambda}{<what param does>}

  \item{scale}{<what param does>}

  \item{...}{<what param does>}
}
\description{
  <full description>
}

