% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/LeastSquaresRegression.R
\name{LeastSquaresRegression}
\alias{LeastSquaresRegression}
\title{<brief desc>}
\usage{
LeastSquaresRegression(X, y, lambda = 0, scale = FALSE, ...)
}
\arguments{
\item{X}{<what param does>}

\item{y}{<what param does>}

\item{lambda}{<what param does>}

\item{scale}{<what param does>}

\item{...}{<what param does>}
}
\description{
<full description>
}

