% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LinearSVM.R
\name{LinearSVM}
\alias{LinearSVM}
\title{Linear SVM Classifier}
\usage{
LinearSVM(X, y, C = 1, method = "Dual", scale = TRUE, eps = 1e-09)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Labels of the observations}

\item{C}{Cost variable}

\item{method}{Estimation procedure c("Dual","Primal","BGD","SGD","Pegasos")}

\item{scale}{Whether a z-transform should be applied (default: TRUE)}

\item{intercept}{Whether an intercept should be added (default: TRUE)}
}
\value{
S4 object of type LinearSVM
}
\description{
Linear SVM Classifier
}

