% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Generics.R, R/KernelLeastSquaresClassifier.R, R/LeastSquaresClassifier.R, R/LinearSVM.R, R/LogisticLossClassifier.R, R/LogisticRegression.R, R/MajorityClassClassifier.R, R/NormalBasedClassifier.R, R/SVM.R, R/SVMlin.R, R/SelfLearning.R, R/TSVMuniversum.R, R/USMLeastSquaresClassifier.R
\docType{methods}
\name{loss}
\alias{line_coefficients}
\alias{loss}
\alias{loss,KernelLeastSquaresClassifier-method}
\alias{loss,LeastSquaresClassifier-method}
\alias{loss,LinearSVM-method}
\alias{loss,LogisticLossClassifier-method}
\alias{loss,LogisticRegression-method}
\alias{loss,MajorityClassClassifier-method}
\alias{loss,NormalBasedClassifier-method}
\alias{loss,SVM-method}
\alias{loss,SVMlin-method}
\alias{loss,SelfLearning-method}
\alias{loss,TSVMuniversvm-method}
\alias{loss,USMLeastSquaresClassifier-method}
\title{Loss of a classifier or regression function}
\usage{
loss(object, ...)

line_coefficients(object, ...)

\S4method{loss}{KernelLeastSquaresClassifier}(object, newdata, y = NULL, ...)

\S4method{loss}{LeastSquaresClassifier}(object, newdata, y = NULL, ...)

\S4method{loss}{LinearSVM}(object, newdata, y = NULL)

\S4method{loss}{LogisticLossClassifier}(object, newdata, y = NULL, ...)

\S4method{loss}{LogisticRegression}(object, newdata, y = NULL)

\S4method{loss}{MajorityClassClassifier}(object, newdata, y = NULL)

\S4method{loss}{NormalBasedClassifier}(object, newdata, y = NULL)

\S4method{loss}{SVM}(object, newdata, y = NULL)

\S4method{loss}{SVMlin}(object, newdata, y = NULL)

\S4method{loss}{SelfLearning}(object, newdata, y = NULL, ...)

\S4method{loss}{TSVMuniversvm}(object, newdata, y = NULL)

\S4method{loss}{USMLeastSquaresClassifier}(object, newdata, y = NULL, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{additional parameters}

\item{newdata}{data.frame object with test data}

\item{y}{Vector with true classes of the test data}

\item{object}{TODO}

\item{y}{Vector with true classes of the test data}

\item{newdata}{data.frame object with test data}

\item{...}{additional parameters}
}
\value{
numeric of the total loss on the test data

numeric of the total loss on the test data
}
\description{
Loss of a classifier or regression function

Loss of a classifier or regression function

Hinge loss on new objects of a trained LinearSVM

Return the MINUS log likelihood on the given dataset

Hinge loss on new objects of a trained SVM

Loss method for SelfLearning Classifier
The loss method delegates prediction to the specific model object
}

