% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{losslogsum}
\alias{losslogsum}
\title{LogsumLoss of a classifier or regression function}
\usage{
losslogsum(object, ...)
}
\arguments{
\item{object}{Classifier or Regression object}

\item{...}{Additional parameters}
}
\description{
LogsumLoss of a classifier or regression function
}

