% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/NormalBasedClassifier.R
\docType{methods}
\name{losslogsum}
\alias{losslogsum}
\alias{losslogsum,NormalBasedClassifier-method}
\title{LogsumLoss of a classifier or regression function}
\usage{
losslogsum(object, ...)

\S4method{losslogsum}{NormalBasedClassifier}(object, newdata, Y, X_u, Y_u)
}
\arguments{
\item{object}{Classifier or Regression object}

\item{...}{Additional parameters}

\item{newdata}{Design matrix of labeled objects}

\item{Y}{label matrix of labeled objects}

\item{X_u}{Design matrix of unlabeled objects}

\item{Y_u}{label matrix of unlabeled objects}
}
\description{
LogsumLoss of a classifier or regression function
}
