% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{losspart}
\alias{losspart}
\title{Loss of a classifier or regression function evaluated on partial labels}
\usage{
losspart(object, ...)
}
\arguments{
\item{object}{TODO}

\item{...}{additional parameters}
}
\value{
numeric of the total loss on the test data
}
\description{
Loss of a classifier or regression function evaluated on partial labels
}

