% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/MajorityClassClassifier.R
\name{MajorityClassClassifier}
\alias{MajorityClassClassifier}
\title{Majority Class Classifier}
\usage{
MajorityClassClassifier(X, y, ...)
}
\arguments{
\item{X}{Design matrix, intercept term is added within the function}

\item{y}{Vector with class assignments}

\item{...}{Unused}
}
\description{
TODO
}

