% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCLinearDiscriminantClassifier.R
\name{MCLinearDiscriminantClassifier}
\alias{MCLinearDiscriminantClassifier}
\title{Moment Constrained Semi-supervised Linear Discriminant Analysis.}
\usage{
MCLinearDiscriminantClassifier(X, y, X_u, method = "closedform",
  prior = NULL, x_center = TRUE, scale = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{method}{character; Currently only "closedform"}

\item{prior}{Matrix (k by 1) of class prior probabilities. If NULL, estimated from data}

\item{x_center}{logical;  Should the features be centered?}

\item{scale}{boolean; Should the features be normalized? (default: FALSE)}
}
\description{
Using an update of the estimated means and covariance proposed in (Loog 2014). 
Using the method="invariant" option, uses the scale invariant parameter update proposed in (Loog 2014), while method="closedform" using the non-scale invariant version from (Loog 2012). The alternative way to minimize this objective function directly has not been implemented yet.
}
\references{
Loog, M., 2012. Semi-supervised linear discriminant analysis using moment constraints. Partially Supervised Learning, LNCS, 7081, pp.32-41.

Loog, M., 2014. Semi-supervised linear discriminant analysis through moment-constraint parameter estimation. Pattern Recognition Letters, 37, pp.24-31.
}
\seealso{
Other RSSL classifiers: \code{\link{MCNearestMeanClassifier}},
  \code{\link{TSVMuniversvm}}
}

