% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCNearestMeanClassifier.R
\name{MCNearestMeanClassifier}
\alias{MCNearestMeanClassifier}
\title{Moment Constrained Semi-supervised Nearest Mean Classifier}
\usage{
MCNearestMeanClassifier(X, y, X_u, method = "closedform", scale = FALSE,
  x_center = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{method}{Estimation procedure: c("closedform","ml")}

\item{scale}{logical; Should the features be normalized? (default: FALSE)}

\item{x_center}{logical;  Should the features be centered?}
}
\value{
S4  object; a list consisting of
\item{means}{the approximation of piel}
\item{prior}{the number of trials}
\item{sigma}{the number of hits}
}
\description{
To fit a true nearest mean classifier, set prior to equal class priors. Based on (Loog 2010).
}
\references{
Loog, M., 2010. Constrained Parameter Estimation for Semi-Supervised Learning: The Case of the Nearest Mean Classifier. In Proceedings of the 2010 European Conference on Machine learning and Knowledge Discovery in Databases. pp. 291-304.
}
\seealso{
Other RSSL classifiers: \code{\link{MCLinearDiscriminantClassifier}},
  \code{\link{TSVMuniversvm}}
}

