% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{missing_labels}
\alias{missing_labels}
\title{Access the true labels for the objects with missing labels when they are stored as an attribute in a data frame}
\usage{
missing_labels(df)
}
\arguments{
\item{df}{data.frame; data.frame with y_true attribute}
}
\description{
Access the true labels for the objects with missing labels when they are stored as an attribute in a data frame
}
\seealso{
Other RSSL utilities: 
\code{\link{LearningCurveSSL}()},
\code{\link{SSLDataFrameToMatrices}()},
\code{\link{add_missinglabels_mar}()},
\code{\link{df_to_matrices}()},
\code{\link{measure_accuracy}()},
\code{\link{split_dataset_ssl}()},
\code{\link{split_random}()},
\code{\link{true_labels}()}
}
\concept{RSSL utilities}
