% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{methods}
\name{plot}
\alias{plot}
\alias{plot,KernelLeastSquaresClassifier,missing-method}
\alias{plot,LeastSquaresClassifier,missing-method}
\alias{plot,SelfLearning,ANY-method}
\alias{show,LeastSquaresClassifier-method}
\title{Plot function}
\usage{
plot(x, y, ...)

\S4method{plot}{KernelLeastSquaresClassifier,missing}(x)

\S4method{show}{LeastSquaresClassifier}(object)

\S4method{plot}{LeastSquaresClassifier,missing}(x)

\S4method{plot}{SelfLearning,ANY}(x, X, y, X_u)
}
\arguments{
\item{x}{Classifier or Regression object}

\item{y}{TODO}

\item{...}{Additional parameters}

\item{X}{Design matrix, intercept term is added within the function}

\item{X_u}{Design matrix unlabeled data}
}
\description{
Plot function

Plot method for SelfLearning Classifier
The plot function plot the label assignments in each iteration
}

