% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearningCurve.R
\name{plot.LearningCurve}
\alias{plot.LearningCurve}
\title{Plot LearningCurve object}
\usage{
\method{plot}{LearningCurve}(x, y, ...)
}
\arguments{
\item{x}{LearningCurve object}

\item{y}{Not used}

\item{...}{Not used}
}
\description{
Plot LearningCurve object
}
