% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/NormalBasedClassifier.R,
%   R/LogisticRegression.R
\docType{methods}
\name{posterior}
\alias{posterior}
\alias{posterior,NormalBasedClassifier-method}
\alias{posterior,LogisticRegression-method}
\title{Class Posteriors of a classifier}
\usage{
posterior(object, ...)

\S4method{posterior}{NormalBasedClassifier}(object, newdata)

\S4method{posterior}{LogisticRegression}(object, newdata)
}
\arguments{
\item{object}{Classifier or Regression object}

\item{...}{Additional parameters}

\item{newdata}{matrix of dataframe of objects to be classified}
}
\description{
Class Posteriors of a classifier
}
