% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Generics.R, R/KernelLeastSquaresClassifier.R, R/LeastSquaresClassifier.R, R/LinearSVM.R, R/LogisticLossClassifier.R, R/LogisticRegression.R, R/NormalBasedClassifier.R, R/SVM.R, R/SelfLearning.R, R/USMLeastSquaresClassifier.R
\docType{methods}
\name{predict}
\alias{decisionvalues,SVM-method}
\alias{predict}
\alias{predict,KernelLeastSquaresClassifier-method}
\alias{predict,LeastSquaresClassifier-method}
\alias{predict,LinearSVM-method}
\alias{predict,LogisticLossClassifier-method}
\alias{predict,LogisticRegression-method}
\alias{predict,NormalBasedClassifier-method}
\alias{predict,SVM-method}
\alias{predict,SelfLearning-method}
\alias{predict,USMLeastSquaresClassifier-method}
\title{Predictions of a classifier or regression function}
\usage{
predict(object, ...)

\S4method{predict}{KernelLeastSquaresClassifier}(object, newdata,
  probs = FALSE, ...)

\S4method{predict}{LeastSquaresClassifier}(object, newdata, probs = FALSE,
  ...)

\S4method{predict}{LinearSVM}(object, newdata)

\S4method{predict}{LogisticLossClassifier}(object, newdata, probs = FALSE)

\S4method{predict}{LogisticRegression}(object, newdata, probs = FALSE)

\S4method{predict}{NormalBasedClassifier}(object, newdata, probs = FALSE)

\S4method{decisionvalues}{SVM}(object, newdata)

\S4method{predict}{SVM}(object, newdata)

\S4method{predict}{SelfLearning}(object, newdata, ...)

\S4method{predict}{USMLeastSquaresClassifier}(object, newdata, probs = FALSE,
  ...)
}
\arguments{
\item{object}{Classification or Regression object}

\item{...}{Additional parameters}

\item{newdata}{new data to predict}

\item{probs}{logical; Whether probabilities should be returned}
}
\description{
Predictions of a classifier or regression function

Predict class of new observations using a LinearSVM

Log likelihood on the training set

Predict class of new observations using a LinearSVM

Predict class of new observations using a LinearSVM

Predict method for SelfLearning Classifier
The Predict Method delegates prediction to the specific model object
}

