% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/scaleMatrix.R
\docType{methods}
\name{predict,scaleMatrix-method}
\alias{predict,scaleMatrix-method}
\title{Predict for matrix scaling inspired by stdize from the PLS package}
\usage{
\S4method{predict}{scaleMatrix}(object, newdata, ...)
}
\description{
Predict for matrix scaling inspired by stdize from the PLS package
}

