% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleMatrix.R
\name{predict,scaleMatrix-method}
\alias{predict,scaleMatrix-method}
\title{Predict for matrix scaling inspired by stdize from the PLS package}
\usage{
\S4method{predict}{scaleMatrix}(object, newdata, ...)
}
\arguments{
\item{object}{scaleMatrix object}

\item{newdata}{data to be scaled}

\item{...}{Not used}
}
\description{
Predict for matrix scaling inspired by stdize from the PLS package
}
