% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCPLDA.R
\name{projection_simplex}
\alias{projection_simplex}
\title{project an n-dim vector y to the simplex Dn}
\usage{
projection_simplex(y)
}
\arguments{
\item{y}{matrix with vectors to be projected onto the simplex}
}
\value{
projection of y onto the simplex
}
\description{
Dn = { x : x n-dim, 1 >= x >= 0, sum(x) = 1}
R translation of Loog's version of Xiaojing Ye's initial implementation.
The algorithm works row-wise
}
\references{
Algorithm is explained as in http://arxiv.org/abs/1101.6081
}
