% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R
\docType{methods}
\name{responsibilities}
\alias{responsibilities}
\title{Responsilibities assigned to the unlabeled objects}
\usage{
responsibilities(object, ...)
}
\arguments{
\item{object}{Classifier; Trained Classifier}

\item{...}{additional parameters}
}
\value{
numeric; responsibilities on the unlabeled objects
}
\description{
Responsilibities assigned to the unlabeled objects
}
