% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Classifier.R, R/LeastSquaresClassifier.R, R/NormalBasedClassifier.R
\docType{methods}
\name{rssl-formatting}
\alias{rssl-formatting}
\alias{show,Classifier-method}
\alias{show,LeastSquaresClassifier-method}
\alias{show,NormalBasedClassifier-method}
\title{Show RSSL classifier}
\usage{
\S4method{show}{Classifier}(object)

\S4method{show}{LeastSquaresClassifier}(object)

\S4method{show}{NormalBasedClassifier}(object)
}
\arguments{
\item{object}{classifier}
}
\description{
Show RSSL classifier

Show the contents of a classifier
}

