% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/Classifier.R,
%   R/NormalBasedClassifier.R, R/scaleMatrix.R
\name{rssl-formatting}
\alias{rssl-formatting}
\alias{show,Classifier-method}
\alias{show,NormalBasedClassifier-method}
\alias{show,scaleMatrix-method}
\title{Show RSSL classifier}
\usage{
\S4method{show}{Classifier}(object)

\S4method{show}{NormalBasedClassifier}(object)

\S4method{show}{scaleMatrix}(object)
}
\arguments{
\item{object}{classifier}
}
\description{
Show RSSL classifier

Show the contents of a classifier
}
