% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Generics.R, R/LeastSquaresClassifier.R,
%   R/NormalBasedClassifier.R, R/LogisticRegression.R, R/GRFClassifier.R,
%   R/KernelLeastSquaresClassifier.R, R/LinearSVM.R, R/LogisticLossClassifier.R,
%   R/MajorityClassClassifier.R, R/SVM.R, R/SelfLearning.R,
%   R/USMLeastSquaresClassifier.R, R/WellSVM.R, R/svmlin.R
\name{rssl-predict}
\alias{rssl-predict}
\alias{predict,LeastSquaresClassifier-method}
\alias{predict,NormalBasedClassifier-method}
\alias{predict,LogisticRegression-method}
\alias{responsibilities,GRFClassifier-method}
\alias{predict,GRFClassifier-method}
\alias{predict,KernelLeastSquaresClassifier-method}
\alias{predict,LinearSVM-method}
\alias{predict,LogisticLossClassifier-method}
\alias{predict,MajorityClassClassifier-method}
\alias{predict,SVM-method}
\alias{predict,SelfLearning-method}
\alias{predict,USMLeastSquaresClassifier-method}
\alias{predict,WellSVM-method}
\alias{decisionvalues,WellSVM-method}
\alias{predict,svmlinClassifier-method}
\title{Predict using RSSL classifier}
\usage{
\S4method{predict}{LeastSquaresClassifier}(object, newdata, ...)

\S4method{predict}{NormalBasedClassifier}(object, newdata)

\S4method{predict}{LogisticRegression}(object, newdata)

\S4method{responsibilities}{GRFClassifier}(object, newdata, ...)

\S4method{predict}{GRFClassifier}(object, newdata = NULL, ...)

\S4method{predict}{KernelLeastSquaresClassifier}(object, newdata, ...)

\S4method{predict}{LinearSVM}(object, newdata)

\S4method{predict}{LogisticLossClassifier}(object, newdata)

\S4method{predict}{MajorityClassClassifier}(object, newdata)

\S4method{predict}{SVM}(object, newdata)

\S4method{predict}{SelfLearning}(object, newdata, ...)

\S4method{predict}{USMLeastSquaresClassifier}(object, newdata, ...)

\S4method{predict}{WellSVM}(object, newdata, ...)

\S4method{decisionvalues}{WellSVM}(object, newdata)

\S4method{predict}{svmlinClassifier}(object, newdata, ...)
}
\arguments{
\item{object}{classifier}

\item{newdata}{objects to generate predictions for}

\item{...}{Other arguments}
}
\description{
Predict using RSSL classifier

For the SelfLearning Classifier the Predict Method delegates prediction to the specific model object
}
