% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4VM.R
\name{S4VM}
\alias{S4VM}
\title{Safe Semi-supervised Support Vector Machine Classifier (S4VM)}
\usage{
S4VM(X, y, X_u, C1 = 100, C2 = 0.1, sample_time = 100, gamma = 0,
  x_center = FALSE, scale = FALSE)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{X_u}{matrix; Design matrix for unlabeled data}

\item{C1}{numeric; regularization parameter for labeled data}

\item{C2}{numeric; regularization parameter for unlabeled data}

\item{sample_time}{numeric; Number of low-density separators that are generated}

\item{gamma}{numeric; width of RBF kernel}

\item{x_center}{logical;  Should the features be centered?}

\item{scale}{logical; Should the features be normalized? (default: FALSE)}
}
\value{
S4 S4VM object with slots:
\item{predictions}{Predictions on the unlabeled objects}
\item{labelings}{Labelings for the different clusters}
}
\description{
R port of the MATLAB implementation of the original authors of the Safe Semi-supervised Support Vector Machine Classifier by Li & Zhou (2011).
}
\details{
The method randomly generates multiple low-density separators (controlled by the sample_time parameter) and merges their predictions by solving a linear programming problem meant to penalize the cost of decreasing the performance of the classifier. S4VM is a bit of a misnomer, since it is a transductive method that only returns predicted labels for the unlabeled objects. The main difference in this implementation compared to the original implementation is the clustering of the low-density separators: in our implementation empty clusters are not dropped during the k-means procedure. In the paper by Li (2011) the features are first normalized to [0,1], which is not automatically done by this function.
}
\references{
Yu-Feng Li and Zhi-Hua Zhou. Towards Making Unlabeled Data Never Hurt. In: Proceedings of the 28th International Conference on Machine Learning (ICML'11), Bellevue, Washington, 2011.
}

