\name{S4VM}
\alias{S4VM}
\title{Safe Semi-Supervised Support Vector Machine Classifier (S4VM)}
\usage{
  S4VM(X, y, X_u)
}
\arguments{
  \item{X}{Design matrix}

  \item{y}{labels}

  \item{X_u}{Design matrix unlabeled data}
}
\description{
  Implementation of the Safe Semi-Supervised Support Vector
  Machine Classifier introduced by Li & Zhou 2011
}

