% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaleMatrix.R
\name{scaleMatrix}
\alias{scaleMatrix}
\title{Matrix centering and scaling}
\usage{
scaleMatrix(x, center = TRUE, scale = TRUE)
}
\arguments{
\item{x}{matrix to be standardized}

\item{center}{TRUE if x should be centered}

\item{scale}{logical; TRUE of x should be scaled by the standard deviation}
}
\description{
This function returns an object with a predict method to center and scale new data. Inspired by stdize from the PLS package
}
