\docType{methods}
\name{show,Classifier-method}
\alias{show,Classifier-method}
\alias{show,NormalBasedClassifier-method}
\title{Show method for Classifier}
\usage{
\S4method{show}{Classifier}(object)

show(object)
}
\arguments{
  \item{object}{object of class NormalBasedClassifier}
}
\description{
Show method for Classifier

Print some properties of classifiers based on classes
modeled as Gaussians.
}

