% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSVM.R
\name{solve_svm}
\alias{solve_svm}
\title{SVM solve.QP implementation}
\usage{
solve_svm(K, y, C = 1)
}
\arguments{
\item{K}{Kernel matrix}

\item{y}{Output vector}

\item{C}{Cost parameter}
}
\description{
SVM solve.QP implementation
}
