% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Evaluate.R
\name{split_dataset_ssl}
\alias{split_dataset_ssl}
\title{Create Train, Test and Unlabeled Set}
\usage{
split_dataset_ssl(X, y, frac_train = 0.8, frac_ssl = 0.8)
}
\arguments{
\item{X}{matrix; Design matrix}

\item{y}{factor; Label vector}

\item{frac_train}{numeric; Fraction of all objects to be used as training objects}

\item{frac_ssl}{numeric; Fraction of training objects to used as unlabeled objects}
}
\description{
Create Train, Test and Unlabeled Set
}
\seealso{
Other RSSL utilities: 
\code{\link{LearningCurveSSL}()},
\code{\link{SSLDataFrameToMatrices}()},
\code{\link{add_missinglabels_mar}()},
\code{\link{df_to_matrices}()},
\code{\link{measure_accuracy}()},
\code{\link{missing_labels}()},
\code{\link{split_random}()},
\code{\link{true_labels}()}
}
\concept{RSSL utilities}
