% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HelperFunctions.R
\name{stderror}
\alias{stderror}
\title{Calculate the standard error of the mean from a vector of numbers}
\usage{
stderror(x)
}
\arguments{
\item{x}{numeric; vector for which to calculate standard error}
}
\description{
Calculate the standard error of the mean from a vector of numbers
}
