% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM.R
\name{SVM}
\alias{SVM}
\title{SVM Classifier}
\usage{
SVM(X, y, C = 1, kernel = NULL, scale = TRUE, intercept = FALSE,
  x_center = TRUE, eps = 1e-09)
}
\arguments{
\item{X}{matrix; Design matrix for labeled data}

\item{y}{factor or integer vector; Label vector}

\item{C}{numeric; Cost variable}

\item{kernel}{kernlab::kernel to use}

\item{scale}{logical; Should the features be normalized? (default: FALSE)}

\item{intercept}{logical; Whether an intercept should be included}

\item{x_center}{logical;  Should the features be centered?}

\item{eps}{Small value to ensure positive definiteness of the matrix in the QP formulation}
}
\value{
S4 object of type SVM
}
\description{
SVM Classifier
}

