% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVM.R
\name{SVM}
\alias{SVM}
\title{SVM Classifier}
\usage{
SVM(X, y, C = 1, method = "Dual", scale = TRUE, intercept = FALSE,
  kernel = NULL, eps = 1e-09)
}
\arguments{
\item{X}{Design matrix}

\item{y}{Labels of the observations}

\item{C}{Cost variable}

\item{method}{Estimation procedure c("Dual","Primal","BGD","SGD","Pegasos")}

\item{scale}{Whether a z-transform should be applied (default: TRUE)}

\item{intercept}{Whether an intercept should be added (default: FALSE)}

\item{kernel}{kernlab::kernel to use}

\item{eps}{Small value to ensure positive definiteness of the matrix in the QP formulation}
}
\value{
S4 object of type SVM
}
\description{
SVM Classifier
}

