% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USMLeastSquaresClassifier.R
\name{USMLeastSquaresClassifier}
\alias{USMLeastSquaresClassifier}
\title{Updated Second Moment Least Squares Classifier}
\usage{
USMLeastSquaresClassifier(X, y, X_u, lambda = 0, intercept = TRUE,
  x_center = FALSE, scale = FALSE, y_scale = FALSE, ...)
}
\arguments{
\item{X}{Design matrix}

\item{y}{labels}

\item{X_u}{Design matrix unlabeled data}

\item{lambda}{numeric; L2 regularization parameter}

\item{intercept}{logical; Whether an intercept should be added to the model}

\item{x_center}{logical; Whether the features should be centered}

\item{scale}{logical; Whether the features should be scaled}

\item{y_scale}{logical; Whether the class vector should be scaled}

\item{...}{Unused}
}
\description{
See for instance \cite{Shaffer1991}. Use all data to estimate E(X'X).
}

