% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CrossValidation.R
\name{xtable.CrossValidation}
\alias{xtable.CrossValidation}
\title{Generate Latex table of cross-validation results}
\usage{
xtable.CrossValidation(object, caption = "", classifier_names = NULL)
}
\arguments{
\item{object}{cross-validation results object}

\item{caption}{character; Caption to be placed over the Latex table}

\item{classifier_names}{character vector; (Shortened) Names of the classifiers}
}
\description{
Generate Latex table of cross-validation results
}

